% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_add_hits,PSMatrix-method}
\alias{.ps_add_hits,PSMatrix-method}
\title{Add Hits Information to a `PSMatrix` Object}
\usage{
\S4method{.ps_add_hits}{PSMatrix}(x, Pos, Strand, Score, Oligo, BG = FALSE, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object to which the hits information will be added.}

\item{Pos}{An integer vector representing the positions of the hits.}

\item{Strand}{A character vector indicating the strand 
of the hits (e.g., "+" or "-").}

\item{Score}{A numeric vector containing the scores of the hits.}

\item{Oligo}{A character vector representing oligonucleotide sequences 
associated with the hits.}

\item{BG}{A logical value indicating whether to treat the 
hits as background (default: `FALSE`).}

\item{withDimnames}{Logical, whether to include dimension names in 
the output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
The updated `PSMatrix` object.
}
\description{
The `.ps_add_hits` method is used to add hit information to a `PSMatrix` 
object, including positional, strand, and score data.
}
