% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_add_hits}
\alias{.ps_add_hits}
\title{Generic Function for Adding Hits}
\usage{
.ps_add_hits(x, ...)
}
\arguments{
\item{x}{An object on which the hits are added.}

\item{...}{Additional arguments passed to methods for `.ps_add_hits`.}
}
\value{
An extended object.
}
\description{
A generic function designed to add hits (matching elements) to an object.
}
