% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.PS_ALPHABET,PSMatrix-method}
\alias{.PS_ALPHABET,PSMatrix-method}
\title{Get the Alphabet from a PSMatrix Object}
\usage{
\S4method{.PS_ALPHABET}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the output, 
if they exist in the object.
Default set to `TRUE`.}
}
\value{
A named vector (integer) representing the alphabet. The names of the 
vector are the nucleotide symbols (e.g., `A`, `C`, `G`, `T`), and the 
values are the corresponding indices (e.g., 1, 2, 3, 4).
}
\description{
This method retrieves the alphabet of nucleotides associated with a 
`PSMatrix` object.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
.PS_ALPHABET(pfm1)

}
