% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_assign_score}
\alias{.ps_assign_score}
\title{Assign a Score to a DNA Sequence}
\usage{
.ps_assign_score(S, x, AB, ncolx)
}
\arguments{
\item{S}{Character. Represents a DNA sequence.}

\item{x}{A `PSMatrix` object containing the position-specific scoring matrix.}

\item{AB}{A named integer vector where names correspond to DNA bases 
and values indicate their respective positions in the scoring matrix.}

\item{ncolx}{description}
}
\value{
A single numeric value representing the score of the input 
   DNA subsequence.
}
\description{
This function computes the score for a given DNA subsequence based on 
a position-specific scoring matrix.
}
