% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_bg_from_table,PSMatrix-method}
\alias{.ps_bg_from_table,PSMatrix-method}
\title{Populates Background Statistics in a `PSMatrix` Object from a Table}
\usage{
\S4method{.ps_bg_from_table}{PSMatrix}(x, short.matrix)
}
\arguments{
\item{x}{An object of class `PSMatrix`.}

\item{short.matrix}{A data.frame or matrix containing background statistics.
The table must have row names matching `ID(x)` and columns `BG_SIZE`,
`BG_MEAN`, `BG_STDEV`.}
}
\value{
The updated `PSMatrix` object.
}
\description{
This method updates a `PSMatrix` object with background statistics 
retrieved from a provided table.
}
\details{
This method uses the `ID(x)` as a key to locate the matching rows in the
`short.matrix` table and updates the following background fields in the 
`PSMatrix` object:
\itemize{
  \item `ps_bg_size`: the size of the background data set.
  \item `ps_bg_avg`: the mean value of the background scores.
  \item `ps_bg_std_dev`: the standard deviation of the background scores.}
  
If no matching row is found in the table, a warning is issued.
}
