% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{.ps_checks}
\alias{.ps_checks}
\title{.ps_checks}
\usage{
.ps_checks(x, pfms, type)
}
\arguments{
\item{x}{Can be a `DNAStringSet` object, a character string that indicates the file path, or 
a `data.frame`.}

\item{pfms}{A list of position frequency matrices representing transcription factors motif,
obtained from the JASPAR database. It is expected to be a `PSMatrixList` object or a `PFMatriList` one.}

\item{type.}{Integer value from 1 to 4, indicating the type of check to be performed. 
If `type == 4`, `pfms` is expected to be a `PSMatrixList` object.
If `type == 3`, `x` is expected to be a `data.frame`.
If `type == 2`, `x` is expected to be a character string indicating a file path. 
If `type == 1` or `type == 4`, `x` is expected to be a `DNAStringSet` object.}
}
\value{
None. The function is used for input validation and will stop with a warning/error 
message if the checks are not met.
}
\description{
Validates input for motif and background sequence analysis.
}
\details{
The function calls `.ps_required_packages()` to ensure that the required packages are loaded. 
Specific checks are performed based on `type` value: 
* `Type == 4`: verifies that `pfms` is a `PSMatrixList` object. Computes the average 
and standard deviation for the background of each motif in pfms. 
If any values are missing (NA), it gives a warning message with the name and ID of the affected motif.
* `Type == 2`: if `x` is a file path (character string), checks if the file is accessible
and search for the `[SHORT TFBS MATRIX]` header in the first line.
* `Type == 1` or `Type == 4`: checks if `x` is a `DNAStringSet` object.
* `Type == 3`: checks if `x` is a `data.frame` containing the BG_SIZE, BG_MEANS, and BG_STDEV
columns. It also checks if all the values in the columns are numeric.
}
