% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{.ps_checks2}
\alias{.ps_checks2}
\title{.ps_checks2()}
\usage{
.ps_checks2(pfms, ...)
}
\arguments{
\item{pfms}{`PSMatrixList` object. If it is not, an error is rised.}

\item{...}{Character string representing a file path. When provided, the function checks if it writable.}
}
\value{
None, it gives errors if the checks are not met.
}
\description{
Provides additional checks on `pfms` parameter and file accessibility for writing.
}
