% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_norm_matrix,PSMatrix-method}
\alias{.ps_norm_matrix,PSMatrix-method}
\title{Normalize a `PSMatrix` object}
\usage{
\S4method{.ps_norm_matrix}{PSMatrix}(x)
}
\arguments{
\item{x}{An object of `PSMatrix` class.}
}
\value{
The updated `PSMatrix` object with normalized columns.
}
\description{
This method normalizes the matrix within a `PSMatrix` object using 
column wise normalization and log-transformation.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
.ps_norm_matrix(pfm1)

}
