% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_norm_score,PSMatrix-method}
\alias{.ps_norm_score,PSMatrix-method}
\title{Normalize the Hit Scores in a `PSMatrix` Object}
\usage{
\S4method{.ps_norm_score}{PSMatrix}(x)
}
\arguments{
\item{x}{An object of class `PSMatrix`.}
}
\value{
A numeric vector of normalized scores.
}
\description{
This method computes normalized hit scores in a `PSMatrix` Object
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
.ps_norm_score(pfm1)

}
