% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_norm_score}
\alias{.ps_norm_score}
\title{Generic Function for Normalizing Scores}
\usage{
.ps_norm_score(x, ...)
}
\arguments{
\item{x}{An object on which the normalization of scores is performed.}

\item{...}{Additional arguments passed to methods for `.ps_norm_score`.}
}
\value{
A numeric vector containing the normalized scores.
}
\description{
A generic function that computes normalized scores for the provided object.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
.ps_norm_score(pfm1)

}
