% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.PS_PSEUDOCOUNT,PSMatrix-method}
\alias{.PS_PSEUDOCOUNT,PSMatrix-method}
\title{Get the Pseudocount Value in a PSMatrix Object}
\usage{
\S4method{.PS_PSEUDOCOUNT}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the output, 
if they exist in the object.
Default set to `TRUE`.}
}
\value{
A numeric values representing the pseudocount added for the calculations.
}
\description{
Get the Pseudocount Value in a PSMatrix Object
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
.PS_PSEUDOCOUNT(pfm1)

}
