% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_scan_s,PSMatrix-method}
\alias{.ps_scan_s,PSMatrix-method}
\title{Scan a Single DNA Sequence Using a PSMatrix Object}
\usage{
\S4method{.ps_scan_s}{PSMatrix}(x, Seq, numx, numx_rc, ncolx, AB)
}
\arguments{
\item{x}{A `PSMatrix` object containing the position-specific scoring matrix.}

\item{Seq}{Character. A single DNA sequence represented as a string.}

\item{numx}{Numeric. Scoring value for the forward strand.}

\item{numx_rc}{Numeric. Scoring value for the reverse strand.}

\item{ncolx}{Integer. Indicates the number of columns}

\item{AB}{A named integer vector where names correspond to DNA bases 
and values indicate their respective positions in the scoring matrix.}
}
\value{
A list containing:
\itemize{
  \item `score`: the highest score obtained.
  \item `strand`: the strand with the highest score.
  \item `pos`: the position of the best match in the sequence. 
  \item `oligo`: the subsequence corresponding to the best match.}
}
\description{
Scan a Single DNA Sequence Using a PSMatrix Object
}
