% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_scan_s}
\alias{.ps_scan_s}
\title{Generic Function for Scanning}
\usage{
.ps_scan_s(x, ...)
}
\arguments{
\item{x}{The object on which the scanning operation is performed.}

\item{...}{Additional arguments passed to methods for `.ps_scan_s`.}
}
\value{
The return value depends on the specific method implementation.
}
\description{
A generic function that performs scanning operations on the provided object.
}
