% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_seq_names,PSMatrix-method}
\alias{.ps_seq_names,PSMatrix-method}
\title{Method for Extracting Sequence Names}
\usage{
\S4method{.ps_seq_names}{PSMatrix}(x, out)
}
\arguments{
\item{x}{An object of class `PSMatrix`.}

\item{out}{The object to which sequence names are assigned.}
}
\value{
An object of class `PSMatrix` to which sequence names are assigned.
}
\description{
This method extract and set sequences name for object of class `PSMatrix`.
}
