% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_seq_names}
\alias{.ps_seq_names}
\title{Retrieve the Sequences Names from an Object}
\usage{
.ps_seq_names(x, out)
}
\arguments{
\item{x}{The object from which sequence names will be processed or retrieved.}

\item{out}{A placeholder}
}
\value{
Depends on the specific method implementation
}
\description{
A generic function designed to process or retrieve sequence names 
from an object.
}
