% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_bg_avg,PSMatrix-method}
\alias{ps_bg_avg,PSMatrix-method}
\title{Get Background Average Score}
\usage{
\S4method{ps_bg_avg}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the
output, if they exist in the object.
Default set to `TRUE`, but this argument is currently unused as
the function returns a single numeric value. It is included for
consistency with other matrix-related methods.}
}
\value{
The stored numeric background average. It is the mean best-hit PWM
   score over the sequences in the background reference set.
}
\description{
Retrieves the background average score stored in a `PSMatrix` object.
This score represents the average binding score of promoter regions
when scanned with a PWM.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_bg_avg(pfm1)

}
