% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_bg_avg<-,PSMatrix-method}
\alias{ps_bg_avg<-,PSMatrix-method}
\title{Set Background Average}
\usage{
\S4method{ps_bg_avg}{PSMatrix}(x) <- value
}
\arguments{
\item{x}{An object of class `PSMatrix`.}

\item{value}{Numeric. The mean background value to be set.}
}
\value{
The modified `PSMatrix` object with the updated background average.
}
\description{
This method specifically sets the background average for an object of
class `PSMatrix`.
}
\details{
The background average is the mean normalized best-hit score across the
background reference set. This method replaces that stored value and then
validates the object.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
`ps_bg_avg<-`(pfm1, value = 0.73473)

}
\seealso{
\code{\link{ps_bg_avg}} to retrieve the background average value,
   and \code{methods::validObject()} for object validation.
}
