% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_bg_avg}
\alias{ps_bg_avg}
\title{Retrieve the Background Average Score for Motif Enrichment Analysis
(Generic Function)}
\usage{
ps_bg_avg(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result): a **numeric value**
  representing the stored background average score.
}
}
\description{
`ps_bg_avg` retrieves the average background score stored in a `PSMatrix`.
}
\details{
The background average score represents the mean motif score computed
over the background sequences. This value is essential for normalizing
motif enrichment results.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_bg_avg(pfm1)

}
