% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_bg_seq_names}
\alias{ps_bg_seq_names}
\title{Retrieve Sequence Names for the Background Dataset (Generic Function)}
\usage{
ps_bg_seq_names(x, ...)
}
\arguments{
\item{x}{An object containing sequence information.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: a **character vector** of sequence
  names corresponding to the analyzed promoter regions.
}
}
\description{
`ps_bg_seq_names` is a **generic function** that extracts the sequence names
or identifiers stored for a full-background reference set.
}
\details{
Methods should be implemented for specific object classes storing sequence
data.


The `ps_bg_seq_names` function is particularly relevant for background
datasets, which include stored motif scan results for every retained
background sequence.

In a PSMatrixList object, the `ps_bg_seq_names` slot is populated only for
full-background matrices. This allows `pscan_full_bg()` to retrieve selected
hits without rescanning DNA.
}
\examples{
full_pfm1_path <- system.file("extdata",
    "full_pfm1.rds",
    package = "PscanR"
)
full_pfm1 <- readRDS(full_pfm1_path)
ps_bg_seq_names(full_pfm1)

}
