% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_bg_size,PSMatrix-method}
\alias{ps_bg_size,PSMatrix-method}
\title{Get Background Size}
\usage{
\S4method{ps_bg_size}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the
output, if they exist in the object.
Included for accessor consistency; it has no effect on this scalar value.}
}
\value{
An integer giving the number of background sequences used for
   background scoring.
}
\description{
Retrieves the number of sequences in the background reference set used to
compute the background statistics in a `PSMatrix` object.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_bg_size(pfm1)

}
