% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_bg_size}
\alias{ps_bg_size}
\title{Retrieve the Background Size Value for Motif Enrichment Analysis
(Generic Function)}
\usage{
ps_bg_size(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result), an **integer value**.
}
}
\description{
`ps_bg_size` is a **generic function** that retrieves the background
size value stored in a `PSMatrix`.
}
\details{
The background size is the number of sequences in the background reference
set used to calculate the stored statistics.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_bg_size(pfm1)

}
