% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_bg_std_dev<-,PSMatrix-method}
\alias{ps_bg_std_dev<-,PSMatrix-method}
\title{Set Background Standard Deviation}
\usage{
\S4method{ps_bg_std_dev}{PSMatrix}(x) <- value
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{value}{Numeric. The value representing the new background
standard deviation.}
}
\value{
The modified `PSMatrix` object with the updated background
   standard deviation.
}
\description{
This method updates the background standard deviation stored in a `PSMatrix`
object.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
`ps_bg_std_dev<-`(pfm1, value = 0.08568925)

}
\seealso{
\code{\link{ps_bg_std_dev}} to retrieve the background
   standard deviation, and \code{methods::validObject()} for validation.
}
