% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_bg_std_dev<-}
\alias{ps_bg_std_dev<-}
\title{Generic Setter Method for Background Standard Deviation Value of an object}
\usage{
ps_bg_std_dev(x, ...) <- value
}
\arguments{
\item{x}{An object for which the background standard deviation
is to be set.}

\item{...}{Additional arguments.}

\item{value}{The value for the background standard deviation.}
}
\value{
An object of the same class of the input with the modified 
   background standard deviation value.
}
\description{
Generic Setter Method for Background Standard Deviation Value of an object
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
`ps_bg_std_dev<-`(pfm1, value = 0.08568925)

}
