% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_bg_std_dev}
\alias{ps_bg_std_dev}
\title{Retrieve the Background Standard Deviation for Motif Enrichment
Analysis (Generic Function)}
\usage{
ps_bg_std_dev(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result): a **numeric value**
  representing the stored background standard deviation value.
}
}
\description{
`ps_bg_std_dev` retrieves the background standard deviation stored in a
`PSMatrix`.
}
\details{
The background standard deviation value quantifies the variability in the
background motif binding scores.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_bg_std_dev(pfm1)

}
