% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_build_bg_from_file}
\alias{ps_build_bg_from_file}
\title{Build background matrices from a file}
\usage{
ps_build_bg_from_file(file, pfms)
}
\arguments{
\item{file}{A character string representing the path for input file. 
This file contains background information for regulatory sequences. Can be generated 
with the `ps_build_bg` function.}

\item{pfms}{A list of position frequency matrices representing transcription factor motif, 
obtained from the JASPAR database.}
}
\value{
A `PSMatrixList` object, containing each motif matrix from `pfms`, background-scored 
using the values provided in `file`. See `ps_build_bg_from_table`.
}
\description{
Generates a background probability profile matrix list by reading an input file.
This function is useful when background information are stored into an external file.
}
\details{
Validates the input with the internal function `.ps_checks`, with `type` set to 2. 
Reads the input file into a `data.frame` with the row names taken from the first column. 
Calls `ps_build_bg_from_table` with the created `data.frame` and `pfms` as input.
}
\examples{
# Load a background information file
file_path <- "../background.txt"

# Load JASPAR motif matrices for vertebrates
opts <- list(collection = "CORE", tax_group = "vertebrates")
J2020 <- getMatrixSet(JASPAR2020, opts)

# Generate the background-scored motif matrices from file
bg_matrices <- ps_build_bg_from_file(file_path, J2020)
}
