% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_fg_avg}
\alias{ps_fg_avg}
\title{Retrieve the Foreground Average Score for Motif Enrichment Analysis
(Generic Function)}
\usage{
ps_fg_avg(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result): a **numeric value**
  representing the stored foreground average score.
}
}
\description{
`ps_fg_avg` retrieves the average foreground score stored in a `PSMatrix`.
}
\details{
The foreground average score represents the mean motif score computed
over the input (foreground) sequences. This value is crucial for
comparing motif occurrences between the foreground and background models.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_fg_avg(pfm1)

}
