% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_fg_size,PSMatrix-method}
\alias{ps_fg_size,PSMatrix-method}
\title{Get Foreground Size}
\usage{
\S4method{ps_fg_size}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the
output, if they exist in the object.
Included for accessor consistency; it has no effect on this scalar value.}
}
\value{
An integer representing the retained foreground size.
}
\description{
Retrieves the number of unique foreground sequences retained after input
cleaning. This includes retained sequences whose best-hit score is `NA`.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_fg_size(pfm1)

}
