% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_fg_size}
\alias{ps_fg_size}
\title{Retrieve the Foreground Size Value for Motif Enrichment Analysis
(Generic Function)}
\usage{
ps_fg_size(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result), an **integer value**.
}
}
\description{
`ps_fg_size` is a **generic function** that retrieves the foreground
size value stored in a `PSMatrix`.
}
\details{
The foreground size is the number of unique sequences retained after input
cleaning, including sequences whose best-hit score is `NA`.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_fg_size(pfm1)

}
