% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_hits_pos}
\alias{ps_hits_pos}
\title{Retrieve Motif Hit Positions (Generic Function)}
\usage{
ps_hits_pos(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: an **integer vector**, where
  names correspond to promoter sequence identifiers, and values represent
  motif hit positions (with an optional shift).
}
}
\description{
`ps_hits_pos` is a **generic function** that retrieves the positions of
motif hits in a given object. These positions indicate where motifs are
located within promoter sequences.
For the `PSMatrix` method, the optional `pos_shift` is the coordinate of the
first sequence base; stored one-based positions are converted accordingly.
}
\details{
Methods should be implemented for specific object classes that store motif
scanning results.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_pos(pfm1)

}
