% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_hits_score,PSMatrix-method}
\alias{ps_hits_score,PSMatrix-method}
\title{Get Hits Score}
\usage{
\S4method{ps_hits_score}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the output, 
if they exist in the object.
Default set to `TRUE`.}
}
\value{
A numeric vector representing the `PSMatrix` hits scores.
}
\description{
Retrieves the hits scores stored in a `PSMatrix` object.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_hits_score(pfm1)

}
