% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_hits_score}
\alias{ps_hits_score}
\title{Retrieve Motif Hit Scores (Generic Function)}
\usage{
ps_hits_score(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: a **named numeric vector**, where
  names correspond to promoter sequence identifiers and values represent
  their respective motif hit scores.
}
}
\description{
`ps_hits_score` is a **generic function** that retrieves the motif hit
scores for each promoter sequence in an object. These scores represent
the binding affinity or enrichment level of promoter sequences when
scanned with a Position Weight Matrix (PWM).
}
\details{
Methods should be implemented for specific object classes that store
motif scanning results.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_score(pfm1)

}
