% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_hits_strand,PSMatrix-method}
\alias{ps_hits_strand,PSMatrix-method}
\title{Get Motif Hit Strand Information}
\usage{
\S4method{ps_hits_strand}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the 
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
A named character vector where names correspond to promoter sequence 
   identifiers, and values represent the strand (`+` or `-`) on which the 
   motif was detected.
}
\description{
Retrieves the strand information (`+` or `-`) of motif hits in a 
`PSMatrix` object. This indicates whether a motif was detected on the 
forward (`+`) or reverse (`-`) strand of the promoter sequence.
The Pscan algorithm scans both the strands of the promoter sequences to 
ensure that no potential binding sites are missed.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_strand(pfm1)

}
