% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_hits_strand_bg}
\alias{ps_hits_strand_bg}
\title{Retrieve Motif Hit Strand Information for the Background Dataset
(Generic Function)}
\usage{
ps_hits_strand_bg(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: a **character vector**, where
  names correspond to promoter sequence identifiers, and values represent
  the strand (`+` or `-`) on which the motif was detected.
}
}
\description{
`ps_hits_strand_bg` is a **generic function** that retrieves the strand
information (`+` or `-`) of motif hits within an object. This indicates
whether a motif was detected on the forward (`+`) or reverse (`-`) strand
of a promoter sequence of the background dataset.
}
\details{
The Pscan algorithm scans both strands of promoter sequences to ensure that
no potential transcription factor binding sites are missed.

Methods should be implemented for specific object classes that store motif
scanning results.


The `ps_hits_strand_bg` function is particularly relevant for background
datasets, which include stored motif scan results for every retained
background sequence.

In a PSMatrixList object, the `ps_hits_strand_bg` slot is populated only
for full-background matrices. This allows `pscan_full_bg()` to retrieve
selected hits without rescanning DNA.
}
\examples{
full_pfm1_path <- system.file("extdata",
    "full_pfm1.rds",
    package = "PscanR"
)
full_pfm1 <- readRDS(full_pfm1_path)
ps_hits_strand_bg(full_pfm1)

}
