% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_hits_z,PSMatrix-method}
\alias{ps_hits_z,PSMatrix-method}
\title{Get Motif Hit Z-score}
\usage{
\S4method{ps_hits_z}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
A named numeric vector where names correspond to promoter sequence
   identifiers and values represent their respective Z-scores.
}
\description{
Computes the Z-Scores for motif hit scores in a `PSMatrix` object.
The Z-score measures how unusual a motif score is compared to background
sequences in the background reference set. Higher Z-scores mean
stronger motif enrichment, suggesting potential regulatory significance.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_z(pfm1)

}
