% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_pvalue}
\alias{ps_pvalue}
\title{Retrieve the Motif-Enrichment P-Value (Generic Function)}
\usage{
ps_pvalue(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result): a **numeric value**
  representing the stored p-value.
}
}
\description{
`ps_pvalue` retrieves the upper-tail p-value stored in a `PSMatrix` after
scanning.
}
\details{
The p-value quantifies the statistical significance of motif
enrichment in the input sequences compared to the background model.
A lower p-value is stronger evidence against the background null model.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_pvalue(pfm1)

}
