% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_score_correlation_map}
\alias{ps_score_correlation_map}
\title{Pscan Score Correlation Heatmap}
\usage{
ps_score_correlation_map(pfms, FDR = 0.01, ...)
}
\arguments{
\item{pfms}{A `PSMatrixList` object containing multiple PWMs and associated 
metadata (foreground and background statistics). Typically is the output 
of `pscan()` function.}

\item{FDR}{Numeric. False Discovery Rate (FDR) threshold to select the TFs
to include in the analysis. The default is `0.01`.}

\item{...}{Additional user defined arguments to customize the heatmap settings, 
such as color palettes or clustering object.}
}
\value{
A heatmap plot showing Z-score correlations for selected 
   transcription factors.
}
\description{
This function generates a heatmap that visualizes the correlation of 
Z-scores for transcription factors (TFs) based on a specified false discovery 
rate threshold. It allows customization of the heatmap's appearance.
}
\details{
The heatmap represents the correlation of motif Z-scores, helping to identify 
clusters of motifs that show similar enrichment patterns across sequences. 
The function performs the following steps:
\itemize{
  \item Extracts the result table and the z-score table from the pscan 
     algorithm result.
  \item Filters the result table based on the given FDR threshold.
  \item Uses the filtered result table to create the z-table for the 
     selected TFs.
  \item Generates the heatmap using the `pheatmap` function, 
     with costumizable settings.}

 Default settings, that can be changed by the users, are: 
 \itemize{
   \item `cluster_rows` and `cluster_cols` set to `TRUE`.
   \item `color` uses a blue-white-red palette.
   \item The main `Pscan Score Correlation Heatmap`.
   \item TF names are showed as column labels.
   }
}
\examples{
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:50]

# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

bg_path <- system.file("extdata", "J2020_hg38_200u_50d_UCSC.psbg.txt", 
                       package = "PscanR")
J2020_PSBG <- ps_retrieve_bg_from_file(bg_path, J2020)

# Execute the Pscan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG, 
                 BPPARAM = BiocParallel::SnowParam(1))
# Use MulticoreParam() for Unix systems (See BiocParallel package).

ps_score_correlation_map(results, FDR = 0.05)

}
