% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_seq_names,PSMatrix-method}
\alias{ps_seq_names,PSMatrix-method}
\title{Get the Sequences Name}
\usage{
\S4method{ps_seq_names}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the 
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
A character vector of names.
}
\description{
Retrieves the names or identifiers of the promoter sequences in a `PSMatrix` 
object. These names correspond to the promoter regions that were scanned by 
the Pscan algorithm.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.RData", package = "PscanR")
load(pfm1_path)
ps_seq_names(pfm1)

}
