% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_seq_names}
\alias{ps_seq_names}
\title{Retrieve Sequence Names (Generic Function)}
\usage{
ps_seq_names(x, ...)
}
\arguments{
\item{x}{An object containing sequence information.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: a **character vector** of sequence
  names corresponding to the analyzed promoter regions.
}
}
\description{
`ps_seq_names` is a **generic function** that extracts the sequence names
or identifiers from an object containing promoter sequence data.
}
\details{
Methods should be implemented for specific object classes storing sequence
data.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_seq_names(pfm1)

}
