% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_transcript_legend,PSMatrixList-method}
\alias{ps_transcript_legend,PSMatrixList-method}
\title{Get the Transcript ID Legend from a PSMatrixList object}
\usage{
\S4method{ps_transcript_legend}{PSMatrixList}(x)
}
\arguments{
\item{x}{An object of class `PSMatrixList`.}
}
\value{
A named character vector containing transcript ID mappings where
   values are the sequence identifiers maintained during the generation of
   the background by the unique() function, while the names are the
   original sequence identifiers for the organism of study.
}
\description{
This method retrieves the value stored in the `transcriptIDLegend` slot of
a `PSMatrixList` object.
}
\examples{
full_pfms_path <- system.file("extdata",
    "full_pfms.rds",
    package = "PscanR"
)
full_pfms <- readRDS(full_pfms_path)
head(ps_transcript_legend(full_pfms))

}
