% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_transcript_legend}
\alias{ps_transcript_legend}
\title{Transcript ID Legend (Generic Function)}
\usage{
ps_transcript_legend(x, ...)
}
\arguments{
\item{x}{A `PSMatrixList` object.}

\item{...}{Additional arguments (not used in this method but included
for extensibility).}
}
\value{
A character vector containing transcript ID mappings.
}
\description{
`ps_transcript_legend` is a **generic function** that retrieves the
`transcriptIDLegend` slot from a `PSMatrixList` object. This slot contains
a character vector mapping transcript IDs used in the dataset.
}
\examples{
full_pfms_path <- system.file("extdata", "full_pfms.rds", package = "PscanR")
full_pfms <- readRDS(full_pfms_path)
head(ps_transcript_legend(full_pfms))

}
