% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_z_table}
\alias{ps_z_table}
\title{Generate a Z-score Matrix of Motif Enrichment per Sequence}
\usage{
ps_z_table(pfms)
}
\arguments{
\item{pfms}{A `PSMatrixList` object containing multiple Position Weight
Matrices and associated metadata (foreground and background statistics).
This object is the output of `pscan()` function.}
}
\value{
A matrix in which each column corresponds to a motif in the `PSMatrixList`
object, and each row to the sequence identifiers. The matrix contains
z-score values for each sequence-motif pair, indicating how much the
computed scores deviate from the expected background distribution.
Higher Z-score values represent stronger motif enrichment.
}
\description{
This function creates a matrix of Z-scores for motif occurrences across
multiple PSMatrix contained in a `PSMatrixList` object. Each row corresponds
to a sequence, and each column represents a motif.
}
\details{
Z-score represents the statistical significance of the alignment scores
for regulatory sequences relative to background expectation. A high Z-score
suggests strong motif enrichment in the foreground compared to
the background.

This function uses example datasets located in the `extdata/` directory for
demonstration purposes only. These files are not part of the core data used
by the function. They can be accessed using `system.file()` as shown in the
examples.
}
\examples{
# Load the promoter sequences for hg38 (Homo sapiens), promoter regions:
# -200 +50 bp in respect to the TSS.
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)[1:10]

# Build a small background from bundled files.
matrix_path <- system.file("extdata", "J2020.rds", package = "PscanR")
matrices <- readRDS(matrix_path)
bg_path <- system.file(
    "extdata", "J2020_hg38_200u_50d_UCSC.psbg.txt",
    package = "PscanR"
)
background <- ps_retrieve_bg_from_file(bg_path, matrices)
motif_ids <- c(
    "MA0506.1", "MA0632.2", "MA0615.1",
    "MA0076.2", "MA0645.1", "MA0838.1"
)
background <- background[motif_ids]

# Execute the Pscan algorithm and view the result table
results <- pscan(prom_seq, background,
    BPPARAM = BiocParallel::SerialParam()
)

ps_z_table(results)

}
