% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_zscore}
\alias{ps_zscore}
\title{Retrieve the Motif-Enrichment Z-Score (Generic Function)}
\usage{
ps_zscore(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result): a **numeric value**
  representing the stored Z-score.
}
}
\description{
`ps_zscore` retrieves the Z-statistic stored in a `PSMatrix` after scanning.
}
\details{
The Z-score represents how significantly a motif is enriched in
the input sequences compared to the background model.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_zscore(pfm1)

}
