% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{pscan_bg}
\alias{pscan_bg}
\title{Extract Pre-Computed Metrics from a PSMatrixList Object}
\usage{
pscan_bg(ID, mega_pfms)
}
\arguments{
\item{ID}{A character vector containing the sequence identifiers of the
sequences to be scanned with the PWMs.}

\item{mega_pfms}{The complete background PSMatrixList generated by
ps_build_bg(). To preserve key metrics, ensure the object is saved using 
save().}
}
\value{
A PSMatrixList object
}
\description{
This function is an optimized version of pscan(), designed to significantly
reduce computational time by extracting relevant metrics from a pre-computed
background PSMatrixList.
}
\details{
If a fully computed PSMatrixList is available, containing scores, positions, 
oligonucleotide sequences, and strand information for all promoter sequences 
of a given organism (scanned using a Position Weight Matrix), this function 
allows users to efficiently retrieve relevant values for specific sequences 
of interest, without having to recalculate them.
}
