% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\docType{class}
\name{PSMatrix-class}
\alias{PSMatrix-class}
\alias{.PSMatrix}
\title{An S4 class to represent a `PFMatrix` object with computed foreground 
(alignment scores) and background statistics}
\description{
The `PSMatrix` class extends the `PFMatrix` class from the `TFBSTools`
package, adding slots for foreground and background statistics, Z-score, 
P-value, and more.
}
\section{Slots}{

\describe{
\item{\code{ps_bg_avg}}{Numeric. The background average value.}

\item{\code{ps_fg_avg}}{Numeric. The foreground average value.}

\item{\code{ps_bg_std_dev}}{Numeric. The standard deviation of the background.}

\item{\code{ps_bg_size}}{Integer. The size of the background dataset.}

\item{\code{ps_fg_size}}{Integer. The size of the foreground dataset.}

\item{\code{ps_hits_pos}}{Integer. Positions of hits.}

\item{\code{ps_hits_strand}}{Character. Strand information for hits.}

\item{\code{ps_hits_score}}{Numeric. Score information of hits.}

\item{\code{ps_hits_oligo}}{Character. Oligonucleotide sequence of hits.}

\item{\code{ps_zscore}}{Numeric. The Z-score value.}

\item{\code{ps_pvalue}}{Numeric. The P-Value.}

\item{\code{ps_seq_names}}{Character. The sequence name.}

\item{\code{.PS_PSEUDOCOUNT}}{Numeric. The pseudocount value used in calculations.}

\item{\code{.PS_ALPHABET}}{Integer. From `1` to `4` to represent the DNA alphabet.}
}}

