% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\docType{class}
\name{PSMatrixList-class}
\alias{PSMatrixList-class}
\alias{.PSMatrixList}
\title{An S4 class to represent a `PSMatrixList` object}
\description{
The `PSMatrixList` class extend the `PFMatrixList` class and represent a
collection of `PSMatrix` object. It allow to aggregate and iterate on 
multiple `PSMatrix` as a cohesive group.
}
\details{
This class extends the `PFMatrixList` class without adding new slots. 
See `PFMatrixList` class documentation.
}
