% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{show,PSMatrix-method}
\alias{show,PSMatrix-method}
\title{Display Details of a `PSMatrix` object}
\usage{
\S4method{show}{PSMatrix}(object)
}
\arguments{
\item{object}{An object of class `PSMatrix`.}
}
\value{
An object of class `PSMatrix` with its details displyed.
}
\description{
This method displays information about a `PSMatrix` object, including 
background and foreground statistics, satndard deviation, background 
and foreground size, z-scores and P-Value.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
show(pfm1)

}
\seealso{
\code{\link{ps_bg_avg}}, \code{\link{ps_fg_avg}}, 
   \code{\link{ps_bg_std_dev}}, \code{\link{ps_bg_size}}, 
   \code{\link{ps_fg_size}}, \code{\link{ps_zscore}}, \code{\link{ps_pvalue}}
}
