% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{transcriptIDLegend}
\alias{transcriptIDLegend}
\title{Retrieve Transcript ID Legend from a PSMatrixList Object (Generic Function)}
\usage{
transcriptIDLegend(x, ...)
}
\arguments{
\item{x}{A `PSMatrixList` object.}

\item{...}{Additional arguments (not used in this method but included for extensibility).}
}
\value{
A character vector containing transcript ID mappings.
}
\description{
`transcriptIDLegend` is a **generic function** that retrieves the 
`transcriptIDLegend` slot from a `PSMatrixList` object. This slot contains 
a character vector mapping transcript IDs used in the dataset.
}
