% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{validPSMatrix}
\alias{validPSMatrix}
\title{Validate a PSMatrix object}
\usage{
validPSMatrix(object)
}
\arguments{
\item{object}{A `PSMatrix` object.}
}
\value{
If all the checks are satisfied, returns `TRUE`. Otherwise, a
string describing the reason of failure.
}
\description{
This function checks if the input `PSMatrix` object properties meet the ones
desired.
}
\details{
The function `validPSMatrix` ensures that the `PSMatrix` object has
correctly formatted background and foreground averages, standard deviations,
and hit-related values.
It performs the following steps:
\itemize{
  \item `ps_bg_avg`,`ps_fg_avg`, and `ps_bg_std_dev` must be of length 1.
  \item Background and foreground averages must be between 0 and 1, and a
  non-missing background standard deviation must be greater than 0 and no
  greater than 1.
  \item Foreground position, strand, score, oligo, and sequence-name vectors
  must have equal lengths.
  \item Stored full-background position, strand, score, oligo, and
  sequence-name vectors must have equal lengths and agree with `ps_bg_size`.}
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
validPSMatrix(pfm1)

}
