% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_bg_from_table}
\alias{.ps_bg_from_table}
\title{Retrieve the background values from a table}
\usage{
.ps_bg_from_table(x, ...)
}
\arguments{
\item{x}{A `data.frame`}

\item{...}{Additional parameters}
}
\description{
Retrieve the background values from a table
}
