% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{.ps_checks}
\alias{.ps_checks}
\title{Check input format}
\usage{
.ps_checks(x, pfms, type)
}
\arguments{
\item{x}{Can be a `DNAStringSet` object, a character string that indicates 
the file path, or a `data.frame`.}

\item{pfms}{A list of position frequency matrices representing transcription 
factors motif, obtained from the JASPAR database. Expected to be a 
`PSMatrixList` or `PFMatrixList` object.}

\item{type}{Integer value from 1 to 4, indicating the type of check to be 
  performed: 

  - `type == 4`: `pfms` is expected to be a `PSMatrixList` object.

  - `type == 3`:`x` is expected to be a `data.frame`.

  - `type == 2`: `x` is expected to be a character string 
    indicating a file path. 
  
  - `type == 1` or `type == 4`: `x` is expected to be a 
    `DNAStringSet` object.}
}
\value{
NULL. The function is used for input validation and will stop with 
a warning/error message if any check fail.
}
\description{
Validates input for motif and background sequence analysis.
}
\details{
The function calls `.ps_required_packages()` to ensure that the required 
packages are loaded. 
Specific checks are performed based on `type` value:
 
* `Type == 4`: Verifies that `pfms` is a `PSMatrixList` object. 
  Computes the average and standard deviation for the background 
  of each motif in `pfms`. If any values are missing (NA), 
  a warning is issued with the name and ID of the affected motif.
  
* `Type == 2`: If `x` is a file path (character string), checks if the file 
  is accessible and search for the `[SHORT TFBS MATRIX]` header in the first 
  line.
  
* `Type == 1` or `Type == 4`: checks if `x` is a `DNAStringSet` object.

* `Type == 3`: Verifies that `x` is a `data.frame` containing the `BG_SIZE`, 
  `BG_MEANS`, and `BG_STDEV` columns. It also checks if all the values in 
  the columns are numeric.
}
\seealso{
[.ps_required_packages()]
}
\keyword{internal}
