% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{.ps_checks2}
\alias{.ps_checks2}
\title{Check input format}
\usage{
.ps_checks2(pfms, ...)
}
\arguments{
\item{pfms}{`PSMatrixList` object. If it is not, an error is rised.}

\item{...}{Can be a character string representing a file path. 
When provided, the function checks if it writable.}
}
\value{
NULL, it gives errors if the checks fail.
}
\description{
Provides additional checks on `pfms` parameter and 
file accessibility for writing.
}
\keyword{internal}
