% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{.ps_norm_matrix}
\alias{.ps_norm_matrix}
\title{Normalize a `PSMatrix` Object}
\usage{
.ps_norm_matrix(x, ...)
}
\arguments{
\item{x}{`PSMatrix` Object}

\item{...}{Additional parameters}
}
\description{
Normalize a `PSMatrix` Object
}
