% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{generate_psmatrixlist_from_background}
\alias{generate_psmatrixlist_from_background}
\title{Generate PSMatrixList from Background data and JASPAR Matrix}
\usage{
generate_psmatrixlist_from_background(
  JASPAR_matrix,
  org,
  prom_reg,
  assembly = character(),
  version = "1",
  destfile = NULL
)
}
\arguments{
\item{JASPAR_matrix}{A character string specifying the JASPAR database 
version. You can choose between 'JASPAR2020', 'JASPAR2022', or 
'JASPAR2024' (non case sensitive).}

\item{org}{A string representing the organism acronym. Accepted values are: 
`hs` (Homo sapiens),
`mm` (Mus musculus), 
`at` (Arabidopsis thaliana),
`sc` (Saccharomyces cerevisiae),
`dm` (Drosophila melanogaster).}

\item{prom_reg}{A numeric vector of two integers as `[upstream, downstream]`
representing the promoter region relative to the TSS. 
Allowed combinations:
\itemize{
  \item 200 base pairs upstream and 50 downstream base pair from the TSS 
  (c(-200, 50))
  \item 450 bp upstream and 50 downstream (c(-450, 50))
  \item 500 bp upstream and 0 downstream (c(-500, 0))
  \item 950 bp upstream and 50 downstream (c(-950, 50))
  \item 1000 bp upstream and 0 downstream (c(-1000, 0))
  }}

\item{assembly}{A string representing the assembly version for Human or 
mouse. For `"hs"` you can choose between `"hg38"` or the latest `"hs1"`.
For `"mm"` you can specify `"mm10"` or `"mm39"`. Default is character().}

\item{version}{A string indicating the version number of the desired 
background. Since annotations may evolve over time, multiple versions 
of the same background could exist. Default is '1', corresponding 
to the first set of backgrounds generated between late 2024 and early 2025. 
Currently, only this initial version is available.}

\item{destfile}{A string indicating the path where the downloaded background 
.txt file should be saved. This tab-separated file contains the matrix 
identifiers, background size, average background score, and standard 
deviation. See \code{\link{ps_retrieve_bg_from_file}} for details on how to 
use this file. Default is NULL, so the file is not saved in user's working 
environment.}
}
\value{
A `PSMatrixList` object with background-scored motif matrices.
}
\description{
This function retrieves background data for a specified organism and 
promoter region, and then fetches a corresponding matrix set from a 
specified JASPAR version. 
It returns a `PSMatrixList` object with background statistics.
}
\details{
The background files are downloaded from the GitHub repository:
\url{https://github.com/dianabetelli/PscanR_backgrounds}
This function automatically fetches the appropriate background file and 
combines it with the specified JASPAR matrix collection to create a 
`PSMatrixList` with background statistics.
}
\examples{
# Note: when running this example, you may see a message indicating that 
# a file is being downloaded. This is expected behavior and not an error — 
# it simply informs you that background data is being retrieved.
bg_matrices <- generate_psmatrixlist_from_background('Jaspar2020', 'hs', 
                                                     c(-200,50), 'hg38')
bg_matrices
bg_matrices[[4]]

}
