% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{generate_psmatrixlist_from_background}
\alias{generate_psmatrixlist_from_background}
\title{Generate PSMatrixList from Background and JASPAR Matrix}
\usage{
generate_psmatrixlist_from_background(JASPAR_matrix, org, prom_reg, assembly)
}
\arguments{
\item{JASPAR_matrix}{A character string specifying the JASPAR database 
version. You can choose between 'JASPAR2020', 'JASPAR2022', or 
'JASPAR2024'}

\item{org}{The accepted values are: `hs` (Homo sapiens), `mm` (Mus musculus), 
`at` (Arabidopsis thaliana), `sc` (Saccharomyces cerevisiae), `dm`
(Drosophila melanogaster).}

\item{prom_reg}{A numeric vector of two integers indicating the range 
of the promoter region. You can choose between:
\itemize{
  \item 200 base pairs upstream and 50 downstream base pair from the TSS 
  (c(-200, 50))
  \item 450 bp upstream and 50 downstream (c(-450, 50))
  \item 500 bp upstream and 0 downstream (c(-500, 0))
  \item 950 bp upstream and 50 downstream (c(-950, 50))
  \item 1000 bp upstream and 0 downstream (c(-1000, 0))
  }}
}
\value{
A `PSMatrixList` object created from the specified background 
file and JASPAR matrix
}
\description{
This function generates an object of class `PSMatrixList` using a specified 
JASPAR matrix, an organism identifier, and a promoter region range. 
It is designed to work with only JASPAR database and UCSC annotations.
}
\examples{
generate_psmatrixlist_from_background('Jaspar2020', 'hs', c(-200,50))

}
