% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_bg_avg<-,PSMatrix-method}
\alias{ps_bg_avg<-,PSMatrix-method}
\title{Set Background Average}
\usage{
\S4method{ps_bg_avg}{PSMatrix}(x) <- value
}
\arguments{
\item{x}{An object of class `PSMatrix`.}

\item{value}{Numeric. The mean background value to be set.}
}
\value{
The modified `PSMatrix` object with the updated background average.
}
\description{
This method specifically set the background average for an object of 
class `PSMatrix`.
}
\details{
The background size is the total number of promoters used to compute background
statistics, such as the average score and standard deviation. This method
allows user to modify this value.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.RData", package = "PscanR")
load(pfm1_path)
`ps_bg_avg<-`(pfm1, value = 0.73473) 

}
\seealso{
\code{\link{ps_bg_avg}} to retrieve the background average value,
   and \code{\link{validObject}} to see which checks are performed on the
   input `PSMatrix` object.
}
