% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_bg_avg<-}
\alias{ps_bg_avg<-}
\title{Generic Setter Method for Background Mean Value of an object}
\usage{
ps_bg_avg(x, ...) <- value
}
\arguments{
\item{x}{An object for which the background average is to be set.}

\item{...}{Additional arguments.}

\item{value}{The value for the background average.}
}
\value{
An object of the same class of the input with the modified 
   background average value.
}
\description{
Generic Setter Method for Background Mean Value of an object
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
`ps_bg_avg<-`(pfm1, value = 0.73473) 

}
