% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_bg_avg}
\alias{ps_bg_avg}
\title{Compute the Background Average Score for Motif Enrichment Analysis 
(Generic Function)}
\usage{
ps_bg_avg(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result): a **numeric value** 
  representing the computed background average score.
  \item If `x` is another supported class, the return format may differ.
}
}
\description{
`ps_bg_avg` is a **generic function** that retrieves the average background 
score for motif enrichment analysis. Methods should be implemented for 
specific object classes that store scan results.
}
\details{
The background average score represents the mean motif score computed 
over the background sequences. This value is essential for normalizing 
motif enrichment results.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_bg_avg(pfm1)

}
