% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_bg_size,PSMatrix-method}
\alias{ps_bg_size,PSMatrix-method}
\title{Get Background Size}
\usage{
\S4method{ps_bg_size}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
An integer value representing the background promoter region size
   used for background scoring.
}
\description{
Retrieves the background promoter region size used to compute the
background statistics in a `PSMatrix` object (e.g., 250L).
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_bg_size(pfm1)

}
