% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_bg_size,PSMatrix-method}
\alias{ps_bg_size,PSMatrix-method}
\title{Get Background Size}
\usage{
\S4method{ps_bg_size}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the 
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
An integer value representing the `PSMatrix` background size.
}
\description{
Retrieves the background dimension stored in a `PSMatrix` object.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_bg_size(pfm1)

}
