% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_bg_size<-,PSMatrix-method}
\alias{ps_bg_size<-,PSMatrix-method}
\title{Setter Method for Background Size in PSMatrix}
\usage{
\S4method{ps_bg_size}{PSMatrix}(x) <- value
}
\arguments{
\item{x}{An object of class `PSMatrix`.}

\item{value}{Numeric. The size background value to be set.}
}
\value{
The modified `PSMatrix` object with the updated background size.
}
\description{
This method specifically set the background size for an object 
of class `PSMatrix`.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
`ps_bg_size<-`(pfm1, value = 25629L)

}
\seealso{
\code{\link{ps_bg_size}} to retrieve the background size, 
   and \code{\link{validObject}} to see which checks are performed on the
   input `PSMatrix` object.
}
