% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_bg_size<-,PSMatrix-method}
\alias{ps_bg_size<-,PSMatrix-method}
\title{Set Background Size in a PSMatrix}
\usage{
\S4method{ps_bg_size}{PSMatrix}(x) <- value
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{value}{An integer representing the new background size.}
}
\value{
The modified `PSMatrix` object with the updated background size.
}
\description{
This method specifically set the background size for a `PSMatrix` object. 
The background size represents the number of promoter regions used in the 
background model.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.RData", package = "PscanR")
load(pfm1_path)
`ps_bg_size<-`(pfm1, value = 25629L)

}
\seealso{
\code{\link{ps_bg_size}} to retrieve the background size, 
   and \code{\link{validObject}} to see which checks are performed on the
   input `PSMatrix` object.
}
