% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_bg_size<-}
\alias{ps_bg_size<-}
\title{Generic Setter Method for Background Size Value of an object}
\usage{
ps_bg_size(x, ...) <- value
}
\arguments{
\item{x}{An object for which the background size value is to be set.}

\item{...}{Additional arguments.}

\item{value}{The value for the background size.}
}
\value{
An object of the same class of the input with the modified 
   background size.
}
\description{
Generic Setter Method for Background Size Value of an object
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
`ps_bg_size<-`(pfm1, value = 25629L)


}
