% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_bg_std_dev,PSMatrix-method}
\alias{ps_bg_std_dev,PSMatrix-method}
\title{Get Background Standard Deviation}
\usage{
\S4method{ps_bg_std_dev}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the 
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
A numeric value representing the standard deviation of the background 
   PWM scores.
}
\description{
Retrieves the background standard deviation of PWM scores stored in a 
`PSMatrix` object. This value quantifies the variability in the background 
motif binding scores.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.RData", package = "PscanR")
load(pfm1_path)
ps_bg_std_dev(pfm1)

}
