% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_bg_std_dev<-,PSMatrix-method}
\alias{ps_bg_std_dev<-,PSMatrix-method}
\title{Setter Method for Background Standard Deviation in PSMatrix}
\usage{
\S4method{ps_bg_std_dev}{PSMatrix}(x) <- value
}
\arguments{
\item{x}{An object of class `PSMatrix`.}

\item{value}{Numeric. The standard deviation background value to be set.}
}
\value{
The modified `PSMatrix` object with the updated background 
   standard deviation.
}
\description{
This method specifically set the background standard deviation
for an object of class `PSMatrix`.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
`ps_bg_std_dev<-`(pfm1, value = 0.08568925)

}
\seealso{
\code{\link{ps_bg_std_dev}} to retrieve the background 
   standard deviation, and \code{\link{validObject}} to see which checks 
   are performed on the input `PSMatrix` object.
}
