% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_bg_std_dev<-}
\alias{ps_bg_std_dev<-}
\title{Set Background Standard Deviation (Generic Function)}
\usage{
ps_bg_std_dev(x, ...) <- value
}
\arguments{
\item{x}{An object to which the background standard deviation will be 
assigned, typically a `PSMatrix` object.}

\item{...}{Additional arguments (not used in the generic function but may be 
used in specific methods).}

\item{value}{A numeric value representing the background standard deviation.}
}
\value{
The updated object with the assigned background standard deviation.
}
\description{
`ps_bg_std_dev<-` is a **generic function** that sets the background 
standard deviation value for a given object.
}
\details{
Methods should be implemented for specific object classes that support 
background standard deviation assignment.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.RData", package = "PscanR")
load(pfm1_path)
`ps_bg_std_dev<-`(pfm1, value = 0.08568925)

}
