% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_build_bg}
\alias{ps_build_bg}
\title{Build background matrices in parallel}
\usage{
ps_build_bg(x, pfms, BPPARAM = bpparam(), BPOPTIONS = bpoptions())
}
\arguments{
\item{x}{A `DNAStringSet` object (see Biostrings package) containing the set 
of regulatory sequences of the organism of study (e.g., gene promoters 
retrieved from a `TxDb` object). These sequences are the target for 
background scanning.}

\item{pfms}{A `PFMatrixlist` object containing position frequency matrices 
representing transcription factor binding preferences. Those are obtained, 
for example, from the JASPAR database.}

\item{BPPARAM}{Parallelization parameter passed to `bplapply` function from 
the `BiocParallel` package. This parameter controls how the parallel
processing is executed.
See `BiocParallel` package for more details.}

\item{BPOPTIONS}{Optional configuration settings passed to `bplapply` 
function from the `BiocParallel` package. This can include additional
options to control the behavior of parallel execution.
See `BiocParallel` documentation for more details.}
}
\value{
A `PSMatrixList` object, containing each motif matrix from `pfms`, 
background-scored against the sequences in `x`.
}
\description{
Generates a background probability profile matrix list using the BiocParallel
framework for parallel execution. It computes background scores for each 
motif matrix from the JASPAR database against a set of regulatory sequences 
(e.g., gene promoters) using the `ps_scan` function.
}
\details{
This function validates input types and remove duplicated sequences
from `x` to avoid redundant computations. It also remove all the sequences 
with an N content above 50% and that have a length different from the one of 
the first sequence, that is considered as a reference. 
The motif matrices are background scored by the `ps_scan` function in parallel.
}
\examples{
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:10]

J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

# Generate the background-scored motif matrices
bg_matrices <- ps_build_bg(prom_seq, J2020, 
                           BPPARAM = BiocParallel::SnowParam(1))
# Use BiocParallel::MulticoreParam() for Unix like systems. 
bg_matrices
}
