% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_build_bg_from_file}
\alias{ps_build_bg_from_file}
\title{Build background matrices from a file}
\usage{
ps_build_bg_from_file(file, pfms)
}
\arguments{
\item{file}{A character string representing the path for input file. 
This file contains background information for regulatory sequences, 
such as size, mean, and standard deviation for background distributions.
The file should be in tabular format, where the first column contains 
the sequence identifiers (row names) and subsequent columns contain the 
background statistics (e.g., `BG_SIZE`, `BG_MEAN`, `BG_STDEV`).
Backgrounds are generated with the `ps_build_bg` function and can be
written to a file with `ps_write_bg_to_file`.}

\item{pfms}{A `PFMatrixList` list of position frequency matrices representing 
transcription factor binding preferences, obtained for example from the JASPAR database. 
These matrices are used to search motif in the background set of regulatory sequences.}
}
\value{
A `PSMatrixList` object, containing each motif matrix from `pfms`, 
background-scored using the values provided in `file`. 
See `ps_build_bg_from_table` for more details.
}
\description{
Import from a file a background score distribution for the input profiles  
computed on a background set of regulatory sequences.
This function reads background information stored into a 
file by `ps_write_bg_to_file`.
}
\details{
This function:

- Validates the input with the internal function `.ps_checks`, 
  with `type` set to 2. 
- Reads the input file into a `data.frame`, with row names taken 
  from the first column. 
- Calls `ps_build_bg_from_table` with the created `data.frame` 
  and `pfms` as input.
}
\examples{
# Load a background information file
file_path <- system.file("extdata", "J2020_hg38_200u_50d_UCSC.psbg.txt", 
                         package = "PscanR")

# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

# Generate the background-scored motif matrices from file
bg_matrices <- ps_build_bg_from_file(file_path, J2020)
bg_matrices
}
\seealso{
\code{\link{ps_build_bg_from_table}}
}
