% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_build_bg_from_table}
\alias{ps_build_bg_from_table}
\title{Build background matrices from table}
\usage{
ps_build_bg_from_table(x, pfms)
}
\arguments{
\item{x}{A `data.frame` containing background parameters for each motif.}

\item{pfms}{A list of position frequency matrices (PFMs) representing transcription factor 
motifs, typically from the JASPAR database. Each element should be coercible to the 
`PSMatrix` class. The number of elements in `pfms` should match the number of row
of `x`}
}
\value{
A `PSMatrixList` object containing each motif matrix from `pfms`, scored with background 
parameters derived from `x`.
}
\description{
Generates a background probability profile matrix list from a background parameter table.
}
\details{
Calls the internal function `.ps_checks()` to validate inputs, converts each
elements of pfms into `PSMatrix` class, and applies the internal function `.ps_bg_from_table`
to each element of `pfms` with the data in `x`. 
A warning is issued if the number of elements in `pfms` doesn't match the number of row of `x`.
}
