% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_build_bg_from_table}
\alias{ps_build_bg_from_table}
\title{Apply background parameters to position frequency matrices}
\usage{
ps_build_bg_from_table(x, pfms)
}
\arguments{
\item{x}{A `data.frame` containing background parameters for each motif.
The data frame should have the following columns:
- `BG_SIZE`: Background size.
- `BG_MEAN`: Mean of the background frequencies.
- `BG_STDEV`: Standard deviation of the background frequencies.
The number of rows in `x` must match the number of elements in `pfms`.}

\item{pfms}{A `PFMatrixList` list of position frequency matrices representing 
transcription factor binding preferences, obtained for example from the 
JASPAR database.
The number of elements in `pfms` should match the number of row of `x`}
}
\value{
A `PSMatrixList` object containing each motif matrix from `pfms`, 
scored with background parameters provided from `x`.
}
\description{
Associates background parameters from a data frame with a list of position 
frequency matrices (PFMs) and returns the list of updated matrices as 
`PSMatrixList` object.
}
\details{
This function: 
\itemize{
   \item Calls the internal function `.ps_checks()` to validate inputs, with 
   `type == 3`.
   \item Converts each elements of pfms into `PSMatrix` class. 
   \item A warning is issued if the number of elements in `pfms` doesn't match 
   the number of row of `x`.
   \item Applies the internal function `.ps_bg_from_table` to each element of 
   `pfms` with the correspponding background data from `x`. 
}
}
\examples{
# create the `data.frame`
background_data <- data.frame(
BG_SIZE = c(500, 450, 480),
BG_MEAN = c(0.3, 0.25, 0.35),
BG_STDEV = c(0.05, 0.07, 0.06)
)

# Retrieve motif matrices for vertebrates from JASPAR2020
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)
J2020_subset <- J2020[1:3] # match the number of rows in `backgound_data`

rownames(background_data) <- c("MA0004.1", "MA0006.1", "MA0019.1")

# Generate background-scored motif matrices
bg_matrices <- ps_build_bg_from_table(background_data, J2020_subset)
bg_matrices
}
