% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_density_plot}
\alias{ps_density_plot}
\title{Density Plot of Motif Hits Along Promoter Regions}
\usage{
ps_density_plot(pfm, shift = 0, st = ps_bg_avg(pfm))
}
\arguments{
\item{pfm}{A `PSMatrix` object. It is a selected matrix from the
`PSMatrixList`, result of `pscan` function.}

\item{shift}{Integer value specifying the positional shift applied to the
hit positions to obtain the position in respect to the TSS.
Default is `0`.}

\item{st}{Score threshold used to filter hits. Can be a numeric value to set
the threshold directly, or a character:
\itemize{
  \item `all`: the threshold is set to `0` (All the hits are evaluated).
  \item `loose`: uses the background average score as threshold.
  \item `strict`: uses the background average score together with the
  background standard deviation as threshold.}
Default is `loose` (background average).}
}
\value{
Invisibly returns `NULL`. The density plot of hit positions along
   the promoter region is drawn as a side effect.
}
\description{
This function creates a density plot representing the distribution of hits
along the promoter sequences based on their position and score for a
specific PWM.
It helps visualize where motif occurrences are concentrated.
}
\details{
The function filters motif hits based on a specified threshold and generates
a density plot to show their distribution. The function includes a vertical
dashed line marking the mode (the most frequent position along the
promoters).

This function uses example datasets located in the `extdata/` directory for
demonstration purposes only. These files are not part of the core data used
by the function. They can be accessed using `system.file()` as shown in the
examples.
}
\examples{
matrix_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
matrix <- readRDS(matrix_path)
ps_density_plot(matrix, shift = -200)

}
