% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_density_plot}
\alias{ps_density_plot}
\title{Pscan Density Plot of Hits along Promoters}
\usage{
ps_density_plot(pfm, shift = 0, st = ps_bg_avg(pfm))
}
\arguments{
\item{pfm}{A Position Frequency Matrix, result of the Pscan algorithm.}

\item{shift}{Numeric value specifying the positional shift applied to the hit 
positions. Default is `0`.}

\item{st}{Score threshold used to filter hits. Can be a numeric value to set 
the threshold directly, or a character:
\itemize{
  \item `all`: the threshold is set to `0` (so, no threshold. All the hits
  are evaluated).
  \item `loose`: uses the background average as threshold.
  \item `strict`: uses the background average plus the background standard 
  deviation as threshold.}
Default is set to loose.}
}
\value{
A density plot showing the distribution of hits along the promoter 
   region.
}
\description{
This function creates a density plot representing the distribution of hits 
along the promoter region based on their position and score.
}
\examples{
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:50]

# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

bg_path <- system.file("extdata", "J2020_hg38_200u_50d_UCSC.psbg.txt", 
                       package = "PscanR")
J2020_PSBG <- ps_retrieve_bg_from_file(bg_path, J2020)

# Execute the Pscan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG, 
                 BPPARAM = BiocParallel::SnowParam(1))
# Use MulticoreParam() for Unix systems (See BiocParallel package).

pfm1 <- results[[1]]
ps_density_plot(pfm1)

}
