% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_fg_avg,PSMatrix-method}
\alias{ps_fg_avg,PSMatrix-method}
\title{Get Foreground Average Score}
\usage{
\S4method{ps_fg_avg}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the output,
if they exist in the object.
Default set to `TRUE`.}
}
\value{
A numeric value representing the foreground average score.
}
\description{
Retrieves the foreground average score stored in a `PSMatrix` object.
}
