% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_fg_avg,PSMatrix-method}
\alias{ps_fg_avg,PSMatrix-method}
\title{Get Foreground Average Score}
\usage{
\S4method{ps_fg_avg}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the 
output, if they exist in the object.
Default set to `TRUE`, but this argument is currently unused as 
the function returns a single numeric value. It is included for 
consistency with other matrix-related methods.}
}
\value{
A numeric value representing the foreground average score. 
   This value is computed as the mean of the PWM scores obtained from 
   scanning a set of promoter regions of genes of interest.
}
\description{
Retrieves the foreground average score stored in a `PSMatrix` object. 
This score represents the average binding score of a set of promoter regions 
of co-expressed or co-regulated genes when scanned with a PWM.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_fg_avg(pfm1)

}
