% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_fg_size,PSMatrix-method}
\alias{ps_fg_size,PSMatrix-method}
\title{Get Foreground Size}
\usage{
\S4method{ps_fg_size}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the output, 
if they exist in the object.
Default set to `TRUE`.}
}
\value{
A numeric value representing the `PSMatrix` foreground size.
}
\description{
Retrieves the foreground dimension stored in a `PSMatrix` object.
}
