% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_fg_size}
\alias{ps_fg_size}
\title{Retrieve the Foreground Size Value for Motif Enrichment Analysis 
(Generic Function)}
\usage{
ps_fg_size(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result), an **integer value**.
  \item If `x` is another supported class, the return format may differ.
}
}
\description{
`ps_fg_size` is a **generic function** that retrieves the foreground 
size value for motif enrichment analysis. Methods should be 
implemented for specific object classes that store scan results.
}
\details{
The foreground size value represent the dimension of the scanned input 
(number of promoter sequences from co-regulated or co-expressed genes 
used as input).
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_fg_size(pfm1)

}
