% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_generics}
\alias{ps_generics}
\alias{ps_zscore}
\alias{ps_pvalue}
\alias{ps_bg_avg}
\alias{ps_fg_avg}
\alias{ps_bg_std_dev}
\alias{ps_bg_size}
\alias{ps_fg_size}
\alias{ps_hits_size}
\alias{ps_hits_score}
\alias{ps_hits_z}
\alias{ps_hits_strand}
\alias{ps_hits_pos}
\alias{ps_hits_oligo}
\alias{ps_hits_table}
\alias{ps_seq_names}
\alias{.PS_PSEUDOCOUNT}
\alias{.PS_ALPHABET}
\alias{ps_scan}
\title{Retrieve or Manipulate Properties of PSMatrix Class}
\usage{
ps_zscore(x, ...)

ps_pvalue(x, ...)

ps_bg_avg(x, ...)

ps_fg_avg(x, ...)

ps_bg_std_dev(x, ...)

ps_bg_size(x, ...)

ps_fg_size(x, ...)

ps_hits_size(x, ...)

ps_hits_score(x, ...)

ps_hits_z(x, ...)

ps_hits_strand(x, ...)

ps_hits_pos(x, ...)

ps_hits_oligo(x, ...)

ps_hits_table(x, ...)

ps_seq_names(x, ...)

.PS_PSEUDOCOUNT(x, ...)

.PS_ALPHABET(x, ...)

ps_scan(x, ...)
}
\arguments{
\item{x}{An object of class `PSMatrix`.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
For `ps_zscore`: a numeric value representing the Z-score.

For `ps_pvalue`: a numeric value representing the P-Value.

For `ps_bg_avg`: a numeric value representing the average background.

For `ps_fg_avg`: a numeric value representing average foreground.

For `ps_bg_std_dev`: a numeric value representing the background 
   standard deviation.

For `ps_bg_size`: an integer value representing the size of the 
   background dataset.

For `ps_fg_size`: an integer value representing the size pf 
   the foreground dataset.

For `ps_hits_size`: an integer value representing the hits size.

For `ps_hits_score`: a numeric vector representing the score for 
   each hit.

For `ps_hits_z`: a numeric vector containing the z scores.

For `ps_hits_strand`: a character vector representing the strand 
   of each hit.

For `ps_hits_pos`: an integer vector representing the position of
   each hit.

For `ps_hits_oligo`: a character vector representing the 
   oligonucleotide sequence of each hit.

For `ps_hits_table`: a `data.frame` of hits.

For `ps_seq_names`: a character vector of sequence names.

For `.PS_PSEUDOCOUNT`: a numeric value representing the pseudocount.

For `.PS_ALPHABET`: integer values.

A `data.frame` of hits.
}
\description{
These generics provide a standardized interface for accessing or manipulating 
various properties of `PSMatrix` objects.
}
\section{Functions}{
\itemize{
\item \code{ps_zscore()}: Retrieve the Z-score from an object

\item \code{ps_pvalue()}: Retrieve the P-Value from an object

\item \code{ps_bg_avg()}: Retrieve the average background value from an object

\item \code{ps_fg_avg()}: Retrieve the average foreground value from an object

\item \code{ps_bg_std_dev()}: Retrieve the background standard deviation from an object

\item \code{ps_bg_size()}: Retrieve the background size from an object

\item \code{ps_fg_size()}: Retrieve the foreground size from an object

\item \code{ps_hits_size()}: Retrieve the hits size from an object

\item \code{ps_hits_score()}: Retrieve the hits score from an object

\item \code{ps_hits_z()}: Retrieve the hits z scores from an object.

\item \code{ps_hits_strand()}: Retrieve the strand of hits from an object

\item \code{ps_hits_pos()}: Retrieve the position of hits from an object

\item \code{ps_hits_oligo()}: Retrieve oligonucleotide sequence of hits from an object

\item \code{ps_hits_table()}: Retrieve the hits table from an object

\item \code{ps_seq_names()}: Retrieve the sequence name from an object

\item \code{.PS_PSEUDOCOUNT()}: Retrieve the pseudocount value

\item \code{.PS_ALPHABET()}: Retrieve the numbers representing the DNA alphabet

\item \code{ps_scan()}: Perform a scan operation on an Object

}}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_zscore(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_pvalue(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_bg_avg(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_fg_avg(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_bg_std_dev(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_bg_size(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_fg_size(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_hits_size(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_hits_score(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_hits_z(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_hits_strand(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_hits_pos(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_hits_table(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_seq_names(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
.PS_PSEUDOCOUNT(pfm1)

pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
.PS_ALPHABET(pfm1)

}
