% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_get_bg_table}
\alias{ps_get_bg_table}
\title{Compute background statistics for position frequency matrices}
\usage{
ps_get_bg_table(pfms)
}
\arguments{
\item{pfms}{A list of position frequency matrices representing the transcription factor motifs.}
}
\value{
A dataframe with one row for each PFM in `pfms`. 
Columns: 
`BG_SIZE`: an integer vector representing the background size for each PFMs. 
`BG_MEAN`: a numeric vector representing the mean of the background frequencies for each PFMs.
`BG_STDEV`: a numeric vector representing the standard deviation of the background frequencies
for each PFMs.
}
\description{
Generates a background statistic table (size, mean, and standard deviation)
from a list of position frequency matrices.
}
\details{
Calls the helper functions `ps_bg_size`, `ps_bg_avg`, and `ps_bg_std_dev` to compute
the statistics. 
In addiction, it also validates the input with `.ps_chechs()` function.
}
