% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_get_bg_table}
\alias{ps_get_bg_table}
\title{Compute background statistics for position frequency matrices}
\usage{
ps_get_bg_table(pfms)
}
\arguments{
\item{pfms}{A list of position frequency matrices representing the 
transcription factor motifs. Each element should be a `PSMatrix` object 
(or should be coercible to `PSMatrix`).}
}
\value{
A dataframe with one row for each PFM in `pfms`.

Columns: 

- `BG_SIZE`: An integer vector representing the background size for each PFMs. 
- `BG_MEAN`: A numeric vector representing the mean of the background 
  frequencies for each PFMs.
- `BG_STDEV`: a numeric vector representing the standard deviation of the 
  background frequencies for each PFMs.
}
\description{
Generates a background statistic table (size, mean, and standard deviation)
from a list of position frequency matrices (PFMs) representing transcription
factor motifs.
}
\details{
This function computes background statistics for each motif matrix in `pfms`.
Before computing the statistics, it validates the input with `.ps_chechs()` 
function.
It calls the helper functions `ps_bg_size`, `ps_bg_avg`, and `ps_bg_std_dev` 
to compute the statistics.
}
\examples{
# Retrieve motif matrices for vertebrates from JASPAR2020
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)
J2020_subset <- J2020[1:3] # match the number of rows in `backgound_data`

# create the `data.frame`
background_data <- data.frame(
BG_SIZE = c(500, 450, 480),
BG_MEAN = c(0.3, 0.25, 0.35),
BG_STDEV = c(0.05, 0.07, 0.06)
)

rownames(background_data) <- c("MA0004.1", "MA0006.1", "MA0019.1")

# Generate background-scored motif matrices
bg_matrices <- ps_build_bg_from_table(background_data, J2020_subset)
bg_table <- ps_get_bg_table(bg_matrices)
bg_table 
}
\seealso{
\code{\link{ps_generics}}
}
