% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_hitpos_map}
\alias{ps_hitpos_map}
\title{Pscan Hits Position Heatmap}
\usage{
ps_hitpos_map(pfms, FDR = 0.01, shift = 0, ...)
}
\arguments{
\item{pfms}{A `PSMatrixList` object containing multiple PWMs and associated 
metadata (foreground and background statistics). Typically is the output 
of `pscan()` function.}

\item{FDR}{Numeric. False Discovery Rate (FDR) threshold to select the transcription 
factors to be included in the analysis. The default is set to `0.01`.}

\item{shift}{Numeric. A value to shift the positions of hits. Default is set to `0`.}

\item{...}{Additional user defined arguments that can be passed to the function 
(e.g., the color palette) to change the default settings.}
}
\description{
This function creates a heatmap visualizing the positional distribution
of hits based on a specified false discovery rate threshold.
}
\details{
The function performs the following steps: 
\itemize{
  \item Extracts the result table from the pscan algorithm result.
  \item Filters the result table based on the given FDR threshold.
  \item talk about the positional map}

Default settings, that can be changed by the users, are:
\itemize{
  \item `cluster_rows` and `cluster_cols` set to `TRUE`.
  \item `color` uses a white-yellow-red palette.
  \item The main `Pscan Hits Position Heatmap`.
  \item }
}
