% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_hits_oligo,PSMatrix-method}
\alias{ps_hits_oligo,PSMatrix-method}
\title{Get Matched Oligonucleotide Sequences}
\usage{
\S4method{ps_hits_oligo}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object, typically the result of the `Pscan` algorithm.}

\item{withDimnames}{Logical, whether to include dimension names in the 
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
A character vector containing the sequences of motif matches 
    (oligonucleotides) in the foreground set.
}
\description{
Retrieves the oligonucleotide sequences that match the motif in the 
foreground sequences, returning a character vector.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_oligo(pfm1)


}
