% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_hits_oligo,PSMatrix-method}
\alias{ps_hits_oligo,PSMatrix-method}
\title{Get The Sequences of Hits}
\usage{
\S4method{ps_hits_oligo}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the 
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
A character containing the sequences of hits of a `PSMatrix` object.
}
\description{
Retrieves the oligonucleotide sequences of hits.
}
