% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_hits_oligo}
\alias{ps_hits_oligo}
\title{Retrieve Matched Oligonucleotide Sequences (Generic Function)}
\usage{
ps_hits_oligo(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: a **named character vector**, where names 
  correspond to sequence identifiers, and values represent the oligonucleotide 
  sequences (subset of the input promoter sequences) matching the motif.
  \item If `x` is another supported class, the return format may differ.
}

For `ps_hits_oligo`: a character vector representing the 
   oligonucleotide sequence of each hit.
}
\description{
`ps_hits_oligo` is a **generic function** that extracts the oligonucleotide 
sequences corresponding to motif hits in a given object.
}
\details{
Methods should be implemented for specific object classes that store motif 
scanning results.
}
