% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_hits_oligo_bg}
\alias{ps_hits_oligo_bg}
\title{Retrieve Matched Oligonucleotide Sequences from a Background Dataset
(Generic Function)}
\usage{
ps_hits_oligo_bg(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: a **character vector**, where
  names correspond to sequence identifiers, and values represent the
  oligonucleotide sequences (subset of the promoter sequences) matching
  the motif.
}
}
\description{
`ps_hits_oligo_bg` is a **generic function** that extracts the
oligonucleotide sequences corresponding to motif hits among a background
dataset in a given object.
}
\details{
Methods should be implemented for specific object classes that store motif
scanning results.


The `ps_hits_oligo_bg` function is particularly relevant for background
datasets, which include stored motif scan results for every retained
background sequence.

In a PSMatrixList object, the `ps_hits_oligo_bg` slot is populated only
for full-background matrices. This allows `pscan_fullBG()` to retrieve
selected hits without rescanning DNA.
}
\examples{
full_pfm1_path <- system.file("extdata",
  "full_pfm1.rds",
  package = "PscanR"
)
full_pfm1 <- readRDS(full_pfm1_path)
ps_hits_oligo_bg(full_pfm1)

}
