% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{ps_hits_score,PSMatrix-method}
\alias{ps_hits_score,PSMatrix-method}
\title{Get Hits Score}
\usage{
\S4method{ps_hits_score}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the 
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
A named numeric vector where names correspond to promoter sequence 
   identifiers and values represent their respective motif hit scores.
}
\description{
Retrieves the motif hit scores for each promoter sequence in a `PSMatrix` 
object.
These scores represent the binding affinity or enrichment level of promoter 
sequences when scanned with a Position Weight Matrix (PWM).
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_score(pfm1)

}
