% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_hits_score_bg}
\alias{ps_hits_score_bg}
\title{Retrieve Motif Hit Scores from a Background Dataset (Generic Function)}
\usage{
ps_hits_score_bg(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: a **named numeric vector**, where
  names correspond to promoter sequence identifiers and values represent
  their respective motif hit scores.
}
}
\description{
ps_hits_score_bg is a generic function designed to extract motif hit scores
from a background dataset. These scores quantify the binding affinity or
enrichment level of promoter sequences when scanned using a Position Weight
Matrix (PWM). This function is specifically useful for analyzing background
matrices, which serve as reference datasets in pscan analyses.
}
\details{
Methods should be implemented for specific object classes that store
motif scanning results.


The `ps_hits_score_bg` function is particularly relevant for background
datasets, which include stored motif scan results for every retained
background sequence. These scores provide a reference for comparison in
motif enrichment analyses, helping to assess the significance of observed
motif occurrences.

In a PSMatrixList object, the `ps_hits_score_bg` slot is populated only for
full-background matrices. This allows `pscan_fullBG()` to retrieve selected
hits without rescanning DNA.
}
\examples{
full_pfm1_path <- system.file("extdata",
  "full_pfm1.rds",
  package = "PscanR"
)
full_pfm1 <- readRDS(full_pfm1_path)
ps_hits_score_bg(full_pfm1)

}
