% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_hits_size,PSMatrix-method}
\alias{ps_hits_size,PSMatrix-method}
\title{Compute Hits Size}
\usage{
\S4method{ps_hits_size}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in the output, 
if they exist in the object.
Default set to `TRUE`.}
}
\value{
An integer representing the total number of motif hits detected in 
   the input promoter sequences.
}
\description{
Retrieves the number of motif hits detected by the Pscan algorithm in a 
`PSMatrix` object. A motif hit represents a significant match between a 
promoter sequence and the Position Weight Matrix (PWM).
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.RData", package = "PscanR")
load(pfm1_path)
ps_hits_size(pfm1)

}
