% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_hits_strand}
\alias{ps_hits_strand}
\title{Retrieve Motif Hit Strand Information (Generic Function)}
\usage{
ps_hits_strand(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: a **character vector**, where 
  names correspond to promoter sequence identifiers, and values represent the 
  strand (`+` or `-`) on which the motif was detected.
  \item If `x` is another supported class, the return format may differ.
}
}
\description{
`ps_hits_strand` is a **generic function** that retrieves the strand 
information (`+` or `-`) of motif hits within an object. This indicates 
whether a motif was detected on the forward (`+`) or reverse (`-`) strand 
of a promoter sequence.
}
\details{
The Pscan algorithm scans both strands of promoter sequences to ensure that 
no potential transcription factor binding sites are missed.

Methods should be implemented for specific object classes that store motif 
scanning results.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_strand(pfm1)

}
