% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_hits_table,PSMatrix-method}
\alias{ps_hits_table,PSMatrix-method}
\title{Retrieve a PSHits Table with Scores, Positions, Strands, and Oligonucleotides}
\usage{
\S4method{ps_hits_table}{PSMatrix}(x, pos_shift = 0L, withDimnames = TRUE)
}
\arguments{
\item{x}{An object of class `PSMatrix`. Should contain the following 
slots:
\itemize{
  \item `ps_hits_score`: a numeric vector of hit scores.
  \item `ps_hits_strand`: a character vector of strand information
  (`-` and `+`).
  \item `ps_hits_oligo`: a character vector of oligo sequences.}}

\item{pos_shift}{Integer. Value for which the position gets shifted.}

\item{withDimnames}{Logical, whether to include dimension names in the output, 
if they exist in the object.
Default set to `TRUE`.}
}
\value{
A `data.frame` ordered by decreasing score value, 
with the following columns:
\itemize{
  \item `SCORE`: the score value of each hit
  \item `POS`: the position value of each hit
  \item `STRAND`: the strand information for each hit
  \item `OLIGO`: the oligo sequence corresponding to each hit} 
Row names correspond to the sequences name.
}
\description{
Retrieve a PSHits Table with Scores, Positions, Strands, and Oligonucleotides
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_hits_table(pfm1)

}
