% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_hits_z}
\alias{ps_hits_z}
\title{Compute Motif Hit Z-Scores (Generic Function)}
\usage{
ps_hits_z(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object: a **named numeric vector**, where 
  names correspond to promoter sequence identifiers and values represent 
  their respective Z-scores.
  \item If `x` is another supported class, the return format may differ.
}
}
\description{
`ps_hits_z` is a **generic function** that computes Z-scores for motif hit 
scores in an object. The Z-score indicates how unusual a motif score is 
compared to background sequences (all promoter sequences in an organism). 
Higher Z-scores suggest stronger motif enrichment, which may indicate 
regulatory significance.
}
\details{
Methods should be implemented for specific object classes that store 
motif scanning results.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_z(pfm1)

}
