% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_pvalue,PSMatrix-method}
\alias{ps_pvalue,PSMatrix-method}
\title{Get P-Value}
\usage{
\S4method{ps_pvalue}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object, typically the result of the `Pscan` algorithm.}

\item{withDimnames}{Logical, whether to include dimension names in the 
output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
A numeric value representing the statistical significance of motif 
    enrichment. Lower values suggest stronger motif enrichment in the 
    foreground set.
}
\description{
Retrieves the P-Value stored in a `PSMatrix` object. The P-value quantifies
the statistical significance of motif enrichment in the input sequences in
respect to the background.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_pvalue(pfm1)

}
