% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_results_table}
\alias{ps_results_table}
\title{Create a Summary Table of PscanR Results}
\usage{
ps_results_table(pfms)
}
\arguments{
\item{pfms}{A `PSMatrixList` object containing multiple PWMs and associated 
metadata (foreground and background statistics). Typically is the output 
of `pscan()` function.}
}
\value{
A data.frame with matrices ordered by increasing P.VALUE and decreasing 
ZSCORE, containig the following columns: 
\itemize{
  \item "NAME": The names of the matrices.
  \item "BG_AVG": The average background score for each PWM.
  \item "BG_STDEV": The standard deviation of the background scores for each 
  PWM.
  \item "FG_AVG": The average foreground score for each PWM.
  \item "ZSCORE": The Z-score for each PWM.
  \item "P.VALUE": The p-value for each PWM.
  \item "FDR": The adjusted p-value (False Discovery Rate) calculated using 
  the Benjamini-Hochberg method.
}
}
\description{
This function generates a table summarizing the results stored in a 
`PSMatrixList` object. It retrieves pre-computed background and foreground 
statistics (the alignment scores between regulatory sequences 
and position weight matrices) for each PWM, calculates adjusted p-values 
(FDR), and returns a table ordered by decreasing `ZSCORE` and 
increasing `P.VALUE`.
}
\examples{

file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:10]

# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

bg_path <- system.file("extdata", "J2020_hg38_200u_50d_UCSC.psbg.txt", 
                       package = "PscanR")
J2020_PSBG <- ps_build_bg_from_file(bg_path, J2020)

# Execute the PScan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG, 
                 BPPARAM = BiocParallel::SnowParam(12))
# Use MulticoreParam() for Unix systems (See BiocParallel package).

table <- ps_results_table(results)

}
\seealso{
\code{\link{ps_generics}}
}
