% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_results_table}
\alias{ps_results_table}
\title{View result in table format}
\usage{
ps_results_table(pfms)
}
\arguments{
\item{pfms}{A `PSMatrixList` object containing multiple PWMs and associated 
metadata (foreground and background statistics). Typically is the output 
of `pscan()` function.}
}
\value{
A data.frame with matrices ordered by decreasing P.VALUE and ZSCORE, 
containig the following columns: 
\itemize{
  \item "NAME": The names of the matrices.
  \item "BG_AVG": The average background score for each PWM.
  \item "BG_STDEV": The standard deviation of the background scores for each 
  PWM.
  \item "FG_AVG": The average foreground score for each PWM.
  \item "ZSCORE": The Z-score for each PWM.
  \item "P.VALUE": The p-value for each PWM.
}
}
\description{
This function generates a table of results from a `PSMatrixList` object, 
ordered by decreasing P.VALUE and ZSCORE. For each PWMs, it computes 
background and #' foreground statistics and returns a table summarizing 
these values.
}
\examples{

# Get the regulatory sequences 
url_target <- "http://159.149.160.88/pscan/sampledata/nfkb100.txt"
target <- read.csv(url_target) 
txdb <- makeTxDbFromUCSC(genome="hg38", tablename="ncbiRefSeqCurated")
seqlevels(txdb) <- seqlevels(txdb)[1:24] # Use only canonical chromosomes
prom_rng <- promoters(txdb, upstream = 500, downstream = 0, use.names = TRUE)
prom_rng$tx_name_clean <- sub("\\\\..*$", "", prom_rng$tx_name)
target_prom_rng <- prom_rng[prom_rng$tx_name_clean \%in\% target[,1]]
prom_seq <- getSeq(x = BSgenome.Hsapiens.UCSC.hg38, target_prom_rng) 

opts <- list(collection = "CORE", tax_group = "vertebrates")
J2020 <- getMatrixSet(JASPAR2020, opts)

J2020_PSBG <- ps_build_bg_from_file(
  "../BG_scripts/J2020_hg38_500u_0d_UCSC.psbg.txt", 
  J2020
)

# Execute the PScan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG)
table <- ps_result_table(result)
View(table)

}
\seealso{
[ps_bg_avg()], [ps_bg_std_dev()], [ps_fg_avg()], [ps_zscore()],
   and [ps_pvalue()]
}
