% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_results_table}
\alias{ps_results_table}
\title{Create a Summary Table of PscanR Results}
\usage{
ps_results_table(pfms)
}
\arguments{
\item{pfms}{A `PSMatrixList` object containing multiple PWMs and associated 
metadata (foreground and background statistics). Typically is the output 
of `pscan()` or `pscan_full_bg()` functions.}
}
\value{
A data.frame with matrices ordered by increasing P.VALUE and decreasing 
ZSCORE, prioritizing significant and strongly enriched motifs.
It contains the following columns: 
\itemize{
  \item "NAME": The matrices identifiers.
  \item "BG_AVG": The average background score of TFBSs for each PWM.
  \item "BG_STDEV": The standard deviation of the background scores of TFBSs
  for each PWM.
  \item "FG_AVG": The average foreground score (the TFBS scores found by 
  scanning the chosen subset of regulatory regions) for each PWM.
  \item "ZSCORE": The Z-score for each PWM computed as 
  (FG_AVG-BG_AVG)/BG_STDEV. It represent a statistical measure for motif 
  enrichment of the scanned sequences in respect to the background.
  \item "P.VALUE": The p-value for each PWM.
  \item "FDR": The adjusted p-value (False Discovery Rate) calculated using 
  the Benjamini-Hochberg correction.
}
}
\description{
This function generates a table summarizing the statistical analysis of motif 
enrichment, stored in a `PSMatrixList` object. It retrieves pre-computed 
background and foreground statistics (the alignment scores between Pscan 
input regulatory sequences and position weight matrices) for each PWM. 
Then it compares motif occurrences in the foreground (e.g., 
coexpressed/coregulated promoter sequences) with the background (e.g., all 
promoters in an organism) using Z-scores, p-values, and FDR correction.
Returns a table ordered by decreasing `ZSCORE` and increasing `P.VALUE`.
}
\details{
This function uses example datasets located in the `extdata/` directory for 
demonstration purposes only. These files are not part of the core data used
by the function. They can be accessed using `system.file()` as shown in the 
examples.
}
\examples{
# Load the promoter sequences for hg38 (Homo sapiens), promoter regions: 
# -200 +50 bp in respect to the TSS. 
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:10]

# Retrieve Background PWMs
J2020_PSBG <- generate_psmatrixlist_from_background('Jaspar2020', 
                                                    'hs', c(-200,50), 'hg38')

# Execute the PScan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG, 
                 BPPARAM = BiocParallel::SnowParam(1))
# Use MulticoreParam() for Unix systems (See BiocParallel package).

ps_results_table(results)

}
