% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_retrieve_bg_from_file}
\alias{ps_retrieve_bg_from_file}
\title{Import background statistics from a file}
\usage{
ps_retrieve_bg_from_file(file, pfms)
}
\arguments{
\item{file}{A character string with the path to the input file. 
This file contains the score distribution statistics for the 
frequency matrices computed on a 
background set of regulatory sequences.
The file should be in tabular format, where the first column contains 
the frequency matrix identifiers (row names) and subsequent columns contain 
the background statistics (e.g., `BG_SIZE`, `BG_MEAN`, `BG_STDEV`).
Background statistics are generated with the `ps_build_bg` function and can 
be written to a file with `ps_write_bg_to_file`.}

\item{pfms}{A `PFMatrixList` object containing position frequency matrices 
representing transcription factor binding preferences, obtained, for example, 
from the JASPAR database. The provided list must correspond to the same 
frequency matrix list used to build the background stored in `file`.}
}
\value{
A `PSMatrixList` object, containing each motif matrix from `pfms`, 
background-scored using the values provided in `file`. 
See `ps_build_bg_from_table` for more details.
}
\description{
Reads and imports background score distribution statistics from a file, 
corresponding to a PSMatrixList object previously computed on a background 
set of regulatory sequences (e.g., the promoters of all transcripts in the 
organism of study, defined by coordinates relative to the TSS) by the 
`ps_build_bg()` function. `ps_retrieve_bg_from_file()` reads the background 
statistics stored into a file by `ps_write_bg_to_file()`.
}
\details{
This function:
\itemize{
   \item Validates the input type: must be a character string.  
   \item Reads the input file into a `data.frame`, with row names taken 
   from the first column. 
   \item Calls `ps_build_bg_from_table` with the created `data.frame` 
   and `pfms` as input.
}

This function uses example datasets located in the `extdata/` directory for 
demonstration purposes only. These files are not part of the core data used
by the function. They can be accessed using `system.file()` as shown in the 
examples.
}
\examples{
# Load a background information file
file_path <- system.file("extdata/BG_files", "J2020_hg38_200u_50d_UCSC.psbg.txt", 
                         package = "PscanR")

# Load the example dataset for JASPAR2020 matrices collection for vertebrates.
J2020_path <- system.file("extdata", "J2020.rds", package = "PscanR")
J2020 <- readRDS(J2020_path)

# Generate the background-scored motif matrices from file
bg_matrices <- ps_retrieve_bg_from_file(file_path, J2020)
bg_matrices
bg_matrices[[2]]
}
\seealso{
\code{\link{ps_build_bg}}, \code{\link{ps_write_bg_to_file}}, 
\code{\link{ps_build_bg_from_table}}

\code{\link{ps_build_bg_from_table}}
}
