% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_scan,PSMatrix-method}
\alias{ps_scan,PSMatrix-method}
\title{Perform a Scan of DNA Sequences Using a `PSMatrix` Object}
\usage{
\S4method{ps_scan}{PSMatrix}(x, seqs, BG = FALSE)
}
\arguments{
\item{x}{An object of `PSMatrix` class.}

\item{seqs}{`A DNAstringSet` object representing the sequences to be
scanned.}

\item{BG}{A logical value indincating whether the scan is for background
sequences. Default is set to `FALSE`.}
}
\value{
The updated `PSMatrix` object with the scanning results.
}
\description{
This method performs a scan of DNA sequences using a `PSMatrix` object 
to identify potential hits based on the matrix score.
}
