% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ps_scan}
\alias{ps_scan}
\title{Perform Motif Scanning (Generic Function)}
\usage{
ps_scan(x, ...)
}
\arguments{
\item{x}{An object containing motif information, typically a `PSMatrix` 
object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
- If `x` is a `PSMatrix` object: a **PSMatrix object** with updated motif hit 
  information, including positions, strands, scores, and matched sequences.

- If `x` belongs to another supported class, the return format may vary.
}
\description{
`ps_scan` is a **generic function** that scans DNA sequences for motif 
occurrences using an object containing motif scoring data. The scan is 
typically performed on both the forward and reverse complement strands to 
detect all potential transcription factor binding sites.
}
\details{
Methods should be implemented for specific object classes that support motif 
scanning.
}
\examples{
BG_matrices <- generate_psmatrixlist_from_background('Jaspar2020', 'hs', 
                                                     c(-200,50),'hg38')
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[25:50]

scanned_result <- ps_scan(BG_matrices[[1]], prom_seq)
scanned_result

}
