% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_score_position_BubbleChart}
\alias{ps_score_position_BubbleChart}
\title{Bubble chart of Pscan Hits Score vs Position}
\usage{
ps_score_position_BubbleChart(pfm, bubble_color = "blue")
}
\arguments{
\item{pfm}{An object of class `PSMatrix`. It must be processed by the PscanR 
algorithm.}
}
\value{
A ggplot object representing a bubble chart. The x-axis corresponds 
to the hits position, the y-axis corresponds to the hits score, and the 
size of the bubbles represents the count of occurrences.
}
\description{
This function visualizes the Bubble Chart to visualize the relationship
between position and score of the identified sites in a `PSMatrix` object.
}
\details{
The function computes the count of occurrences for each combination of 
PS hits score and position from the input `PSMatrix` and visualizes 
this data using a bubble chart. The user can modify the color of bubbles. 
Default is `blue`.
}
\examples{
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:50]

# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

bg_path <- system.file("extdata", "J2020_hg38_200u_50d_UCSC.psbg.txt", 
                       package = "PscanR")
J2020_PSBG <- ps_build_bg_from_file(bg_path, J2020)

# Execute the Pscan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG, 
                 BPPARAM = BiocParallel::SnowParam(1))
# Use MulticoreParam() for Unix systems (See BiocParallel package).

pfm1 <- results[[1]]
ps_score_position_BubbleChart(pfm1)

}
