% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_score_position_BubbleChart}
\alias{ps_score_position_BubbleChart}
\title{Bubble chart of Pscan Motif Score vs Position}
\usage{
ps_score_position_BubbleChart(pfm, bubble_color = "blue")
}
\arguments{
\item{pfm}{A `PSMatrix` object. It must be processed by the PscanR 
algorithm.}

\item{bubble_color}{A character string specifying the color of the bubbles 
(default: `"blue"`).}
}
\value{
A ggplot2 bubble chart. The x-axis corresponds to the position of 
motif hits along the promoter, the y-axis corresponds to the score of motif 
hits, and the size of the bubbles represents the count of occurrences for 
each score-position combination.
}
\description{
This function creates a Bubble Chart to visualize the relationship
between position and score of the identified sites along the promoter 
sequences in a `PSMatrix` object.
}
\details{
The function computes the count of occurrences for each combination of 
PS hits score and position from the input `PSMatrix` and visualizes 
this data using a bubble chart. The user can modify the color of bubbles. 
Default is `blue`.
}
\examples{
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:25]

# Retrieve Background PWMs
J2020_PSBG <- generate_psmatrixlist_from_background('Jaspar2020', 
                                                    'hs', c(-200,50), 'hg38')

# Execute the Pscan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG, 
                 BPPARAM = BiocParallel::SnowParam(1))
# Use MulticoreParam() for Unix systems (See BiocParallel package).

pfm1 <- results[[1]]
ps_score_position_BubbleChart(pfm1)

}
