% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_write_bg_to_file}
\alias{ps_write_bg_to_file}
\title{Write background statistics to a file}
\usage{
ps_write_bg_to_file(pfms, file)
}
\arguments{
\item{pfms}{A list of position frequency matrices representing transcription factor motif, 
obtained from the JASPAR database.}

\item{file}{A character string specifing the path where the result should be saved.}
}
\value{
None. It saves the given background statistics in a file.
}
\description{
This function computes background statistics for each element of a list of position frequency
matrices, and write the result in a specified file.
}
\details{
It validates the input by calling the heper function `ps.checks2()`.
By calling ps_get_bg_table, it computes the background statistics. 
A header is added to the file (`[SHORT TFBS MATRIX`]).
}
