% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Background.R
\name{ps_write_bg_to_file}
\alias{ps_write_bg_to_file}
\title{Save background statistics from a `PSMatrixList` object to a file}
\usage{
ps_write_bg_to_file(pfms, file)
}
\arguments{
\item{pfms}{A `PSMatrixList` object of Position Weight Matrices 
representing transcription factor binding preferences, obtained for example 
from the JASPAR database, containing background statistics computed by 
scanning each `PSMatrix` against the set of promoter regions of all 
transcripts in the organism of study. The promoter regions analyzed may 
vary. Each element should be a `PSMatrix` object or coercible to `PSMatrix`.}

\item{file}{A character string specifying the path to the output file where 
the background statistics should be saved.}
}
\value{
None. It saves the given background statistics to the specified file
in a tab-delimited format.
}
\description{
Saves background statistics (such as size, mean, and standard deviation) 
for each Position Weight Matrix in a `PSMatrixList` object to a specified 
file.
}
\details{
This function retrieves the background statistics (size, mean, and 
standard deviation) using `ps_get_bg_table()` from the input `PSMatrixList` 
object, after having validated the inputs. Then, it writes the result to a 
specified file. 
A header is added to the file (`[SHORT TFBS MATRIX`]).
}
\examples{
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)
# File path to save the result
file_path <- "J2020_hg38_bg_stats.txt"

PSM1 <- PSMatrix(
  pfm = J2020[[1]],
  ps_bg_avg = 0.25,        
  ps_fg_avg = 0.5,         
  ps_bg_std_dev = 0.05,    
  ps_bg_size = 250L        
  )
PSM2 <- PSMatrix(
  pfm = J2020[[2]],
  ps_bg_avg = 0.25,        
  ps_fg_avg = 0.5,         
  ps_bg_std_dev = 0.05,    
  ps_bg_size = 250L        
  )
  
PSMatrixList_J2020 <- PSMatrixList(PSM1, PSM2)

# ps_write_bg_to_file(PSMatrixList_J2020, file_path)

}
\seealso{
\code{\link{ps_get_bg_table}}
}
