% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_z_table}
\alias{ps_z_table}
\title{Generates Z-score table}
\usage{
ps_z_table(pfms)
}
\arguments{
\item{pfms}{A `PSMatrixList` object containing multiple PWMs and associated 
metadata (foreground and background statistics). Typically is the output 
of `pscan()` function.}
}
\value{
A matrix in which each column corresponds to a motif in the `PSMatrixList`
object, and each row to the sequence identifiers.
}
\description{
This function calculates the z-score for each motif in a `PSMatrixList` 
object and organizes the result into a matrix.
}
\examples{
# Get the regulatory sequences 
target <- read.csv("../PscanR/example_file_nfkb100.txt", header = FALSE) 
txdb <- makeTxDbFromUCSC(genome="hg38", tablename="ncbiRefSeqCurated")
seqlevels(txdb) <- seqlevels(txdb)[1:24] # Use only canonical chromosomes
prom_rng <- promoters(txdb, upstream = 500, downstream = 0, use.names = TRUE)
prom_rng$tx_name_clean <- sub("\\\\..*$", "", prom_rng$tx_name)
target_prom_rng <- prom_rng[prom_rng$tx_name_clean \%in\% target[,1]]
prom_seq <- getSeq(x = BSgenome.Hsapiens.UCSC.hg38, target_prom_rng) 

opts <- list(collection = "CORE", tax_group = "vertebrates")
J2020 <- getMatrixSet(JASPAR2020, opts)

J2020_PSBG <- ps_build_bg_from_file(
  "../BG_scripts/J2020_hg38_500u_0d_UCSC.psbg.txt", 
  J2020
)

# Execute the PScan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG, BPPARAM = MultiCoreParam(24))
# Use `SnowParam` on windows
z_score <- ps_z_table(result)
View(z_score)

}
