% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_z_table}
\alias{ps_z_table}
\title{Generate a Z-score Table for Motif Hits}
\usage{
ps_z_table(pfms)
}
\arguments{
\item{pfms}{A `PSMatrixList` object containing multiple Position Weight 
Matrices and associated metadata (foreground and background statistics). 
This object is the output of `pscan()` function.}
}
\value{
A matrix in which each column corresponds to a motif in the `PSMatrixList`
object, and each row to the sequence identifiers. The cell values in the 
matrix are Z-scores, indicating the statistical significance of the alignment 
between each sequence and the motif.
}
\description{
This function creates a matrix of Z-scores for motif occurrences across 
multiple PSMatrix contained in a `PSMatrixList` object. Each row corresponds 
to a sequence, and each column represents a motif.
}
\details{
The Z-Score represents the statistical significance of the alignment scores 
for regulatory sequences relative to background expectation. A high Z-score 
suggests strong motif enrichment in the foreground compared to 
the background.
}
\examples{

# The generation of the example might take few minutes 

file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:10]

# Retrieve Background PWMs
J2020_PSBG <- generate_psmatrixlist_from_background('Jaspar2020', 
                                                    'hs', c(-200,50), 'hg38')

# Execute the Pscan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG, 
                 BPPARAM = BiocParallel::SnowParam(1))
# Use MulticoreParam() for Unix systems (See BiocParallel package).

ps_z_table(results)

}
