% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{ps_z_table}
\alias{ps_z_table}
\title{Generate a Z-score Table for Motifs}
\usage{
ps_z_table(pfms)
}
\arguments{
\item{pfms}{A `PSMatrixList` object containing multiple Position Weight 
Matrices and associated metadata (foreground and background statistics). 
This object is the output of `pscan()` function.}
}
\value{
A matrix in which each column corresponds to a motif in the `PSMatrixList`
object, and each row to the sequence identifiers. The cell values in the 
matrix are Z-scores, indicating the statistical significance of the alignment 
between each sequence and the motif.
}
\description{
This function calculates the Z-score for each motif in a `PSMatrixList` 
object and organizes the results into a matrix. The Z-Score represents
the statistical significance of the alignment scores for regulatory 
sequences relative to background expectation.
}
\examples{
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:10]

# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

bg_path <- system.file("extdata", "J2020_hg38_200u_50d_UCSC.psbg.txt", 
                       package = "PscanR")
J2020_PSBG <- ps_retrieve_bg_from_file(bg_path, J2020)

# Execute the Pscan algorithm and view the result table
results <- pscan(prom_seq, J2020_PSBG, 
                 BPPARAM = BiocParallel::SnowParam(1))
# Use MulticoreParam() for Unix systems (See BiocParallel package).

ps_z_table(results)

}
