% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_zscore,PSMatrix-method}
\alias{ps_zscore,PSMatrix-method}
\title{Get Z-Score}
\usage{
\S4method{ps_zscore}{PSMatrix}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{A `PSMatrix` object.}

\item{withDimnames}{Logical, whether to include dimension names in 
the output, if they exist in the object.
Default set to `TRUE`.}
}
\value{
A numeric value representing the `PSMatrix` Z-score.
}
\description{
Retrieves the Z-Score stored in a `PSMatrix` object.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
ps_zscore(pfm1)

}
