% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{ps_zscore}
\alias{ps_zscore}
\title{Compute the Z-score for motif enrichment analysis (Generic Function)}
\usage{
ps_zscore(x, ...)
}
\arguments{
\item{x}{An object containing motif scan results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
\itemize{
  \item If `x` is a `PSMatrix` object (Pscan result): a **numeric value** 
  representing the computed Z-score.
  \item If `x` is another supported class, the return format may differ.
}
}
\description{
`ps_zscore` is a **generic function** that computes the Z-score 
for motif enrichment analysis. Methods should be implemented for 
specific object classes that store scan results.
}
\details{
The Z-score represents how significantly a motif is enriched in 
the input sequences compared to the background model.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.RData", package = "PscanR")
load(pfm1_path)
ps_zscore(pfm1)

}
