% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{pscan}
\alias{pscan}
\title{Executes the Pscan algorithm on a set of regulatory sequences.}
\usage{
pscan(x, pfms, BPPARAM = bpparam(), BPOPTIONS = bpoptions())
}
\arguments{
\item{x}{A `DNASetString` object containing the set of regulatory sequences 
from co-regulated or co-expressed genes (i.e. a set of gene promoters). 
(see Biostrings package).}

\item{pfms}{An object of PSMatrixList class containing PWMs and background 
values. See ps_build_bg, ps_build_bg_from_file, ps_build_bg_from_table for 
how to create PSMatrixList objects.}

\item{BPPARAM}{The BPPARAM used by bplapply. See BiocParallel.
This argument is passed to `BiocParallel::bplapply`
The default is `bpparam()`.}

\item{BPOPTIONS}{The BPOPTIONS used by bplapply. See BiocParallel.
This argument is passed to `BiocParallel::bplapply`. 
The default is `bpoptions()`.}
}
\value{
A `PSMatrixList` object in which the foreground value have been computed 
for each sequence in `X` based on the position weight matrices in `pfms`.
}
\description{
This function computes the foreground values of a set of gene promoters (`X`)
using the PScan algorithm. Foreground values are computed by applying the 
PWM values stored in `pfms`.
}
\examples{

file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:10]

# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

bg_path <- system.file("extdata", "J2020_hg38_200u_50d_UCSC.psbg.txt", 
                       package = "PscanR")
J2020_PSBG <- ps_build_bg_from_file(bg_path, J2020)

# Execute the PScan algorithm
results <- pscan(prom_seq, J2020_PSBG, 
                 BPPARAM = BiocParallel::MulticoreParam(1))
results

}
