% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{pscan}
\alias{pscan}
\title{Executes the Pscan algorithm on a set of sequences.}
\usage{
pscan(x, pfms, BPPARAM = bpparam(), BPOPTIONS = bpoptions())
}
\arguments{
\item{x}{The set of regulatory sequences from co-regulated or co-expressed genes (i.e.
a set of gene promoters). This must be a DNAStringSet object (see Biostrings package)}

\item{pfms}{An object of PSMatrixList class containing PWMs and background values. See 
ps_build_bg, ps_build_bg_from_file, ps_build_bg_from_table for how to create
PSMatrixList objects.}

\item{BPPARAM}{The BPPARAM used by bplapply. See BiocParallel.}

\item{BPOPTIONS}{The BPOPTIONS used by bplapply. See BiocParallel.}
}
\value{
Returns pfms with foreground values computed on x.
}
\description{
Executes the Pscan algorithm on a set of sequences.
}
