% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scan_and_post_processing.R
\name{pscan_fullBG}
\alias{pscan_fullBG}
\title{Extract Pre-Computed Metrics from a PSMatrixList Object}
\usage{
pscan_fullBG(ID, full_pfms)
}
\arguments{
\item{ID}{A character vector containing sequence identifiers (transcript 
IDs) corresponding to the sequences of interest. These should match 
those used when the background was originally computed 
(e.g., RefSeq transcript IDs).}

\item{full_pfms}{A `PSMatrixList` object generated using 
\code{\link{ps_build_bg}} with the `fullBG = TRUE` flag. This object 
must contain a complete set of pre-computed scores and metadata for all 
promoter sequences of the reference organism.}
}
\value{
A `PSMatrixList` object in which the alignment scores and related 
   metrics have been retrieved from `full_pfms` for each sequence in `ID`, 
   based on the background data.
}
\description{
This function is a faster alternative to \code{\link{pscan}}, that works
only if a full background PSMatrixList object has already been computed. 
Instead of re-scanning input sequences, it retrieves pre-computed alignment 
scores, TFBS positions, strands, and binding oligos for a selected subset of 
sequences from the complete background.
}
\details{
This is particularly useful when analyzing different subsets of sequences on
on the same background, reducing computation time.


Internally, the function matches the provided `ID` values against the 
`transcriptIDLegend` slot of the full background PSMatrixList. 
The transcriptIDLegend slot was introduced to handle cases where duplicate 
promoter sequences are removed using unique() during background generation. 
Since in this case only one representative sequence name is retained,  
the mapping vector ensures that if a user provides a transcript ID 
corresponding to a removed duplicate, it can still be correctly associated 
with the retained sequence used in the background.
Transcript versions (e.g., ".1", ".2") are ignored for matching purposes.

Sequences that were excluded during preprocessing (e.g., due to high 'N' 
content or length mismatch) will be excluded with a warning printed.
}
\examples{
# load a full background PSMatrixList 
full_bg_path <- system.file('extdata', 'full_pfms.rds', package = 'PscanR')
full_pfms <- readRDS(full_bg_path)

IDs <- c('NM_031921.6', 'NM_001005484.2', 'NR_047525.1', 'NR_029639.1',
         'NR_036051.1', 'NR_029834.1', 'NM_001029885.2', 'NR_148357.1',
         'NR_148960.1', 'NM_001130413.4')

results <- pscan_fullBG(IDs, full_pfms)
ps_results_table(results)

}
