% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{PSMatrix}
\alias{PSMatrix}
\title{Create a PSMatrix object}
\usage{
PSMatrix(
  pfm,
  ps_bg_avg = as.numeric(NA),
  ps_fg_avg = as.numeric(NA),
  ps_bg_std_dev = as.numeric(NA),
  ps_bg_size = as.integer(NA),
  .PS_PSEUDOCOUNT = 0.01,
  ...
)
}
\arguments{
\item{pfm}{An object of class `PFMatrix` from the `TFBSTools` package 
representing a position frequency matrix.}

\item{ps_bg_avg}{Numeric. The background average value, default = `NA`.}

\item{ps_fg_avg}{Numeric. The foreground average value, default = `NA`.}

\item{ps_bg_std_dev}{Numeric. The standard deviation of the background, 
default = `NA`.}

\item{ps_bg_size}{Integer. The size of the background dataset, 
default = `NA`.}

\item{.PS_PSEUDOCOUNT}{Numeric. The pseudocount to add to avoid division
by zero. Default = `0.01`}

\item{...}{Addictional arguments to pass to downstream function.}
}
\value{
An object of `PSMatrix` class.
}
\description{
This function creates an instance of the `PSMatrix` class by initializing its
slots with default values.
}
