% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{PSMatrix}
\alias{PSMatrix}
\title{Create a PSMatrix object}
\usage{
PSMatrix(
  pfm,
  ps_bg_avg = as.numeric(NA),
  ps_fg_avg = as.numeric(NA),
  ps_bg_std_dev = as.numeric(NA),
  ps_bg_size = as.integer(NA),
  .PS_PSEUDOCOUNT = 0.01,
  ...
)
}
\arguments{
\item{pfm}{An object of class `PFMatrix` from the `TFBSTools` package 
representing a position frequency matrix.}

\item{ps_bg_avg}{Numeric. The background average value, default = `NA`.}

\item{ps_fg_avg}{Numeric. The foreground average value, default = `NA`.}

\item{ps_bg_std_dev}{Numeric. The standard deviation of the background, 
default = `NA`.}

\item{ps_bg_size}{Integer. The size of the background dataset, 
default = `NA`.}

\item{.PS_PSEUDOCOUNT}{Numeric. The pseudocount to add to avoid division
by zero. Default = `0.01`}

\item{...}{Additional arguments to pass to downstream function.}

\item{from}{A `PFMatrix` object to be converted to a `PSMatrix` one.}
}
\value{
An object of `PSMatrix` class.

A `PSMatrix` object created from the input `PFMatrix` object.
}
\description{
This function creates an instance of the `PSMatrix` class by initializing its
slots with default values.
}
\examples{

# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

result <- PSMatrix(
  pfm = J2020[[1]],
  ps_bg_avg = 0.25,        
  ps_fg_avg = 0.5,         
  ps_bg_std_dev = 0.05,    
  ps_bg_size = 250L        
  )
print(result)
}
