% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-PSMatrix.R
\name{PSMatrix}
\alias{PSMatrix}
\alias{setAs,PFMatrix,PSMatrix-method}
\title{An S4 Class for a Single Positiont-Specific Matrix}
\usage{
PSMatrix(
  pfm,
  ps_bg_avg = as.numeric(NA),
  ps_fg_avg = as.numeric(NA),
  ps_bg_std_dev = as.numeric(NA),
  ps_bg_size = as.integer(NA),
  .PS_PSEUDOCOUNT = 0.01,
  ...
)
}
\arguments{
\item{pfm}{An object of class `PFMatrix` from the `TFBSTools` package, 
typically a position frequency matrix (PFM) representing the motif for 
which all the organism's promoter regions will be scanned.}

\item{ps_bg_avg}{Numeric. The background average value, default = `NA`.
It is the average binding score of promoter regions when scanned with a 
PWM.}

\item{ps_fg_avg}{Numeric. The foreground average value, default = `NA`.
Represents the average binding score of a set of promoter regions of 
co-expressed or co-regulated genes when scanned with a PWM.}

\item{ps_bg_std_dev}{Numeric. The background standard deviation of PWM 
scores, default = `NA`.}

\item{ps_bg_size}{Integer. The size of the background promoter region, 
default = `NA`.}

\item{.PS_PSEUDOCOUNT}{Numeric. The Pseudo-Count to add to avoid division
by zero. Default = `0.01`}

\item{...}{Additional arguments passed to other methods or used in the 
initialization.}
}
\value{
A `PSMatrix` object containing the provided statistics, initialized 
   with default or provided values.

A `PSMatrix` object created from the input `PFMatrix` object.
}
\description{
This function creates a `PSMatrix` object, which contains a single 
Position Frequency Matrix (PFM) and associated metadata related to promoter 
sequence motif hits, such as background and foreground averages, standard 
deviation, size, z-scores, and other related information.
}
\details{
When a `PSMatrix` object is created or modified, it is automatically 
validated using an internal function. The validation ensures that the 
background and foreground statistics are properly formatted, and that motif 
hit-related vectors have consistent lengths.
}
\examples{

# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rds", package = "PscanR")
J2020 <- readRDS(J2020_path)

result <- PSMatrix(
  pfm = J2020[[1]],
  ps_bg_avg = 0.25,        
  ps_fg_avg = 0.5,         
  ps_bg_std_dev = 0.05,    
  ps_bg_size = 250L        
  )
print(result)
}
