% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{PSMatrixList}
\alias{PSMatrixList}
\alias{setAs,PFMatrixList,PSMatrixList-method}
\title{Create a `PSMatrixList` object}
\usage{
PSMatrixList(..., transcriptIDLegend = character(), use.names = TRUE)
}
\arguments{
\item{...}{Objects of class `PSMatrix` to include in the list.}

\item{use.names}{Logical. Assert whether to use names from the input objects.
Default = `TRUE`}
}
\value{
A `PSMatrixList` object, which is a list containing `PSMatrix` 
   objects. Each element in the list corresponds to a `PSMatrix` object 
   provided as input.

A `PSMatrixList` object, which is a list containing the `PSMatrix` 
   objects converted from the `PFMatrixList`.
}
\description{
This function creates a `PSMatrixList` object, which is a container for 
managing multiple `PSMatrix` object.

This method converts a `PFMatrixList` object into a `PSMatrixList` object by 
applying the conversion for each `PFMatrix` in the list to a `PSMatrix`. 
The resulting `PSMatrixList` is created by combining the converted `PSMatrix` 
objects into a list and returning the new object.
}
\examples{
# Load JASPAR motif matrices for vertebrates
J2020_path <- system.file("extdata", "J2020.rda", package = "PscanR")
load(J2020_path)

PSM1 <- PSMatrix(
  pfm = J2020[[1]],
  ps_bg_avg = 0.25,        
  ps_fg_avg = 0.5,         
  ps_bg_std_dev = 0.05,    
  ps_bg_size = 250L        
  )
PSM2 <- PSMatrix(
  pfm = J2020[[2]],
  ps_bg_avg = 0.25,        
  ps_fg_avg = 0.5,         
  ps_bg_std_dev = 0.05,    
  ps_bg_size = 250L        
  )
result <- PSMatrixList(PSM1, PSM2)
ps_results_table(result)
}
