% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{PSMatrixList}
\alias{PSMatrixList}
\title{Create a `PFMatrixList` object}
\usage{
PSMatrixList(..., use.names = TRUE)
}
\arguments{
\item{...}{Objects of class `PSMatrix` to include in teh list.}

\item{use.names}{Logical. Assert whether to use names from the input objects.
Default = `TRUE`}
}
\value{
An object of class `PSMatrixList`.
}
\description{
This function creates a `PSMatrixList` object, which is a container for 
managing multiple `PSMatrix` object.
}
