% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{show,PSMatrix-method}
\alias{show,PSMatrix-method}
\title{Display Details of a `PSMatrix` object}
\usage{
\S4method{show}{PSMatrix}(object)
}
\arguments{
\item{object}{An object of class `PSMatrix`.}
}
\value{
Prints a summary of the `PSMatrix` object, including:
\itemize{
  \item Pscan Background Average
  \item Pscan Foreground (the sample) Average
  \item Pscan Background Standard Deviation
  \item Pscan Background Size
  \item Pscan Foreground (the sample) Size
  \item Pscan Z-score
  \item Pscan p-value (Z-test)
}
}
\description{
This method displays a summary of a `PSMatrix` object, including 
background and foreground statistics, standard deviation, background 
and foreground size, z-scores and P-Value.
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.RData", package = "PscanR")
load(pfm1_path)
show(pfm1)

}
\seealso{
\code{\link{ps_bg_avg}}, \code{\link{ps_fg_avg}}, 
   \code{\link{ps_bg_std_dev}}, \code{\link{ps_bg_size}}, 
   \code{\link{ps_fg_size}}, \code{\link{ps_zscore}}, \code{\link{ps_pvalue}}
}
