% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PSMatrix.R
\name{transcriptIDLegend,PSMatrixList-method}
\alias{transcriptIDLegend,PSMatrixList-method}
\title{Get the Transcript ID Legend from a PSMatrixList object}
\usage{
\S4method{transcriptIDLegend}{PSMatrixList}(x)
}
\arguments{
\item{x}{An object of class `PSMatrixList`.}
}
\value{
A named character vector containing transcript ID mappings where 
   values are the sequence identifiers maintained during the generation of 
   the background by the unique() function, while the names are the 
   original sequence identifiers for the organism of study.
}
\description{
This method retrieves the value stored in the `transcriptIDLegend` slot of
a `PSMatrixList` object.
}
