% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatrix_class.R
\name{validPSMatrix}
\alias{validPSMatrix}
\title{Validate a `PSMatrix` object}
\usage{
validPSMatrix(object)
}
\arguments{
\item{object}{An object of class `PSMatrix`.}
}
\value{
If all the checks are satisfied, returns `TRUE`. Otherwise, a
string describing the reason of failure.
}
\description{
This function checks if the input `PSMatrix` object properties meet the ones
desired.
}
\details{
The function `validPSMatrix` validates many properties of a `PSMatrix` 
object, including: 
\itemize{
  \item `ps_bg_avg`,`ps_fg_avg`, and `ps_bg_std_dev` must be of length 1.
  \item `ps_bg_avg` and `ps_bg_std_dev` value must be between 0 and 1, 
  excluding 0 for `ps_bg_std_dev`.
  \item the length of `ps_hits_pos`, `ps_hits_strand`, and `ps_hits_score` 
  vectors must be equal.}
}
\examples{
pfm1_path <- system.file("extdata", "pfm1.rda", package = "PscanR")
load(pfm1_path)
validPSMatrix(pfm1)

}
